Enum boltzmann_solver::common::Statistic

pub enum Statistic {
    FermiDirac,
    BoseEinstein,
    MaxwellBoltzmann,
}

Variants
FermiDiracBoseEinsteinMaxwellBoltzmann
Methods

impl Statistic

pub fn phase_space(&self, e: f64, mu: f64, beta: f64) -> f64

pub fn number_density(&self, mass: f64, mu: f64, beta: f64) -> f64

Return number density for a particle following the specified statistic.

\[ n = \frac{1}{2 \pi^2} \int_{m}^{\infty} f(u, \mu, \beta) u \sqrt{u^2 - m^2} \dd u \]

Implementation Details

If an analytic closed form is available for the integral, it will be preferred over the numerical integration. Failing that, the semi-infinite integral is numerical evaluated. In order to make the calculation tractable on a computer, the change of variables

\[ u \to m + \frac{t}{1 - t}, \qquad \dd u \to \frac{\dd t}{(t - 1)^2} \]

is used such that the bounds of the integral over \(t\) are \([0, 1]\).

pub fn massless_number_density(&self, mu: f64, beta: f64) -> f64

Return number density for a massless particle following the specified statistic.

\[ n = \frac{1}{2 \pi^2} \int_{0}^{\infty} f(u, \mu, \beta) u^2 \dd u \]

This is theoretically equivalent to setting mass = 0 in [Statistic::number_density]; however, as there are analytic closed forms of the above integral for all statistics, this method will be much faster and more precise.