Module boltzmann_solver::constants

Predefined constants.

Although computing some of these values is generally very fast (e.g. PI_2 == PI.powi(2)), it will be slightly faster to use pre-defined constant which, in certain performance-sensitive functions, can result in an appreciable performance improvement.

Constants

PI
\(\pi\)

PI_1
\(\pi\), with a different name to follow the convention for PI_n.

PI_2
\(\pi^2\)

PI_M1
\(\pi^{-1}\)

PI_M2
\(\pi^{-2}\)

PLANCK_MASS
Planck mass, \(M_{\text{Pl}} = \sqrt{\hbar c / G}\), in units of GeV / \(c^2\).

REDUCED_PLANCK_MASS
Reduced Planck mass, \(m_{\text{Pl}} = \sqrt{\hbar c / 8 \pi G}\), in units of GeV / \(c^2\).

ZETA_3
The Riemann zeta function evaluated at 3.