Enum boltzmann_solver::Statistic

pub enum Statistic {
    FermiDirac,
    BoseEinstein,
    MaxwellBoltzmann,
    MaxwellJuttner,
}

The statistics which describe the distribution of particles over energy states. Both Fermi–Dirac and Bose–Einstein quantum statistics are implemented, as well as the classical Maxwell–Boltzmann statistic.

Variants
FermiDirac

Fermi–Dirac statistic describing half-integer-spin particles:

\begin{equation} f_{\textsc{FD}} = \frac{1}{\exp[(E - \mu) \beta] + 1}. \end{equation}

BoseEinstein

Bose–Einstein statistic describing integer-spin particles:

\begin{equation} f_{\textsc{BE}} = \frac{1}{\exp[(E - \mu) \beta] - 1}. \end{equation}

MaxwellBoltzmann

Maxwell–Boltzmann statistic describing classical particles:

\begin{equation} f_{\textsc{MB}} = \exp[-(E - \mu) \beta]. \end{equation}

MaxwellJuttner

Maxwell–Jüttner statistic describing relativistic classical particles:

\begin{equation} f_{\textsc{MJ}} = \frac{E \beta \sqrt{E^2 - m^2}}{m K_2(m \beta)} \exp[- E \beta]. \end{equation}

Methods

impl Statistic

pub fn phase_space(&self, e: f64, m: f64, mu: f64, beta: f64) -> f64

Evaluate the phase space distribution, \(f\) as defined above.

pub fn number_density(&self, mass: f64, mu: f64, beta: f64) -> f64

Return number density for a particle following the specified statistic.

\begin{equation} n = \frac{1}{2 \pi^2} \int_{m}^{\infty} f_{i} u \sqrt{u^2 - m^2} \dd u \end{equation}

Implementation Details

If an analytic closed form is available for the integral, it will be preferred over the numerical integration. Failing that, the semi-infinite integral is numerical evaluated. In order to make the calculation tractable on a computer, the change of variables

\begin{equation} u \to m + \frac{t}{1 - t}, \qquad \dd u \to \frac{\dd t}{(t - 1)^2} \end{equation}

is used such that the bounds of the integral over \(t\) are \([0, 1]\).

pub fn massless_number_density(&self, mu: f64, beta: f64) -> f64

Return number density for a massless particle following the specified statistic.

\begin{equation} n = \frac{1}{2 \pi^2} \int_{0}^{\infty} f_{i} u^2 \dd u \end{equation}

This is theoretically equivalent to setting mass = 0 in [Statistic::number_density]; however, as there are analytic closed forms of the above integral for all statistics, this method will be much faster and more precise.