Module boltzmann_solver::model::interaction

Common trait and implementations for interactions.

Structs

DisplayError

Error used when failing to find the particle within the model.

FastInteractionResult

Result from a fast interaction.

FourParticle

Three particle interaction, all determined from the underlying squared amplitude.

PartialWidth

Partial width associated with a single specific interaction.

Particles

List of particles involved in the interaction.

RateDensity

Rate density associate with an interaction.

ThreeParticle

Three particle interaction, all determined from the underlying squared amplitude.

Traits

Interaction

Generic interaction between particles.

Functions

fix_equilibrium

Adjust dn and/or dna for to account for: