Struct boltzmann_solver::model::interaction::FourParticle

pub struct FourParticle<M> { /* private fields */ }

Three particle interaction, all determined from the underlying squared amplitude.

Implementations

impl<M> FourParticle<M>

pub fn new<F>(
squared_amplitude: F,
p1: isize,
p2: isize,
p3: isize,
p4: isize
) -> Self where
F: Fn(&M, f64, f64, f64) -> f64 + Sync + 'static,

Create a new four-particle interaction.

The interaction will result in the annihilation of p1 and p2, and the creation of p3 and p4. Particles are indicated by positive integers and the corresponding antiparticles are indicates by negative integers.

The squared amplitude is defined of the form

\abs{\scM(p_1 p_2 \leftrightarrow p_3 p_4)}^2(s, t, u)

where $s$, $t$ and $u$ are the usual Mandelstam variables. Crossing symmetry is then used in order to compute the other processes.

The three-body decays are not currently computed.

pub fn new_all<F>(
squared_amplitude: F,
p1: isize,
p2: isize,
p3: isize,
p4: isize
) -> Vec<Self> where
F: Fn(&M, f64, f64, f64) -> f64 + Sync + Copy + 'static,

Create a set of related four-particle interactions.

This functions creates all three possible interactions from the same underlying squared amplitude, related through crossing symmetry:

\\begin{aligned}
  p_1 p_2 \leftrightarrow p_3 p_4, \\\\
  p_1 \overline{p_3} \leftrightarrow \overline{p_2} p_4, \\\\
  p_1 \overline{p_4} \leftrightarrow \overline{p_2} p_3. \\\\
\\end{aligned}

The squared amplitude is defined of the form:

\abs{\scM(p_1 p_2 \leftrightarrow p_3 p_4)}^2(s, t, u)

where $s$, $t$ and $u$ are the usual Mandelstam variables. Crossing symmetry is then used in order to compute the other processes.

pub fn set_asymmetry<F>(&mut self, asymmetry: F) -> &Self where
F: Fn(&M, f64, f64, f64) -> f64 + Sync + 'static,

Specify the asymmetry between this process and its $\CP$-conjugate.

This asymmetry is specified in terms of the asymmetry in the squared amplitudes:

\delta \abs{\scM}^2
  \defeq \abs{\scM(p_1 p_2 \to p_3 p_4)}^2 - \abs{\scM(\overline{p_1} \overline{p_2} \to \overline{p_3} \overline{p_4})}^2
  = \abs{\scM(p_1 p_2 \to p_3 p_4)}^2 - \abs{\scM(p_3 p_4 \to p_1 p_2)}^2

This asymmetry is subsequently used to compute the asymmetric interaction rate given by Interaction::delta_gamma.

pub fn enable_width(self, v: bool) -> Self

Adjust whether this interaction will calculate decay widths.

pub fn enable_gamma(self, v: bool) -> Self

Adjust whether this interaction will calculate decay widths.

impl<M: Model> FourParticle<M>

pub fn count_heavy(&self, context: &Context<'_, M>) -> usize

Count the number of heavy particles