Trait boltzmann_solver::model::interaction::Interaction

pub trait Interaction<M: Model> {
    fn particles(&self) -> &InteractionParticles;

    fn particles_idx(&self) -> &InteractionParticleIndices;

    fn particles_sign(&self) -> &InteractionParticleSigns;

    fn gamma(&self, c: &Context<M>) -> f64;

    fn width(&self, _: &Context<M>) -> Option<PartialWidth> { ... }

    fn calculate_rate(&self, gamma: f64, c: &Context<M>) -> RateDensity { ... }

    fn adjust_rate_overshoot(
        &self, 
        rate: RateDensity, 
        c: &Context<M>
    ) -> RateDensity { ... }

    fn change(
        &self, 
        dn: Array1<f64>, 
        dna: Array1<f64>, 
        c: &Context<M>
    ) -> (Array1<f64>, Array1<f64>) { ... }
}

Generic interaction between particles.

Required methods

fn particles(&self) -> &InteractionParticles

Return the particles involved in this interaction

fn particles_idx(&self) -> &InteractionParticleIndices

Return the particles involved in this interaction, not distinguishing between particles an antiparticles.

fn particles_sign(&self) -> &InteractionParticleSigns

Return the sign of the particles, with +1 for a particle and -1 for an antiparticle.

fn gamma(&self, c: &Context<M>) -> f64

Calculate the reaction rate density.

This is returned as a vector for each possible configuration.

Loading content...
Provided methods

fn width(&self, _: &Context<M>) -> Option<PartialWidth>

Calculate the decay width associated with a particular interaction.

There may not be a result if the decay is not kinematically allowed, or not relevant.

The default implementation simply returns None and must be implemented manually.

fn calculate_rate(&self, gamma: f64, c: &Context<M>) -> RateDensity

Calculate the interaction rates density from the interaction rate density by taking into account the number densities prefactors.

fn adjust_rate_overshoot(
&self,
rate: RateDensity,
c: &Context<M>
) -> RateDensity

Adjust the backward and forward rates such that they do not overshoot the equilibrium number densities.

fn change(
&self,
dn: Array1<f64>,
dna: Array1<f64>,
c: &Context<M>
) -> (Array1<f64>, Array1<f64>)

Add this interaction to the dn array.

The changes in dn should contain only the effect from the integrated collision operator and not take into account the normalization to inverse-temperature evolution nor the dilation factor from expansion of the Universe. These factors are handled separately and automatically.

This function automatically adjusts the rate calculated by Interaction::calculate_rate so that overshooting is avoided.

Loading content...