Struct boltzmann_solver::model::EmptyModel

pub struct EmptyModel { /* fields omitted */ }

Empty model with no particles or interactions.

As the 0 index is reserved for a particle that is it’s own antiparticle, this is the only ‘particle’ included within the model.

Implementations

impl EmptyModel

pub fn push_particle(&mut self, p: Particle)

Push the specified particle to the end of the list of particles within the model.

Note that the particle comes with a default none particle at index 0, thus the index of particles added with this start from 1.

pub fn extend_particles<I>(&mut self, iter: I) where
I: IntoIterator<Item = Particle>,

Extend the particles within the model given the specified iterator.

Note that the particle comes with a default none particle at index 0, thus the index of particles added with this start from 1.

pub fn push_interaction<I>(&mut self, interaction: I) where
I: Interaction<Self> + Sync + 'static,

Push the specified interaction into the model.

pub fn extend_interactions<I>(&mut self, iter: I) where
I: IntoIterator,
<I as IntoIterator>::Item: Interaction<Self> + Sync + 'static,

Extend the particles within the model given the specified iterator.

Note that the particle comes with a default none particle at index 0, thus the index of particles added with this start from 1.