Trait boltzmann_solver::model::Model

pub trait Model where
    Self: Sized, {

    fn zero() -> Self;

    fn set_beta(&mut self, beta: f64);

    fn get_beta(&self) -> f64;

    fn entropy_dof(&self, beta: f64) -> f64;

    fn particles(&self) -> &[Particle];

    fn particles_mut(&mut self) -> &mut [Particle];

    fn static_particle_idx<S: AsRef<str>>(
        name: S, 
        i: usize
    ) -> Result<usize, (S, usize)>;

    fn hubble_rate(&self, beta: f64) -> f64 { ... }

    fn len_particles(&self) -> usize { ... }

    fn particle_idx<S: AsRef<str>>(
        &self, 
        name: S, 
        i: usize
    ) -> Result<usize, (S, usize)> { ... }

    fn particle_num<S: AsRef<str>>(
        &self, 
        name: S, 
        i: usize
    ) -> Result<isize, (S, usize)> { ... }

    fn static_particle_num<S: AsRef<str>>(
        name: S, 
        i: usize
    ) -> Result<isize, (S, usize)> { ... }

    fn particle_name(&self, i: isize) -> Result<String, isize> { ... }

    fn particle<S: AsRef<str>>(&self, name: S, i: usize) -> &Particle { ... }

    fn particle_mut(&mut self, name: &str, i: usize) -> &mut Particle { ... }

    fn as_context(&self) -> Context<'_, Self> { ... }

}

Contains all the information relevant to a particular model, including masses, widths and couplings. All these attributes can be dependent on the inverse temperature $\beta$.

Required methods

fn zero() -> Self

Instantiate a new instance of the model at 0 temperature ($\beta = \infty$).

This is implemented separately so that the computational peculiarities of dealing with $\beta = \infty$ can be avoided.

fn set_beta(&mut self, beta: f64)

Update the model to be valid at the given inverse temperature beta.

The inverse temperature is always strictly positive, and should never be infinite. The zero temperature case should be obtained from Model::zero.

fn get_beta(&self) -> f64

Return the current value of beta for the model

fn entropy_dof(&self, beta: f64) -> f64

Return the effective degrees of freedom contributing to the entropy density of the Universe at the specified inverse temperature.

fn particles(&self) -> &[Particle]

Return a list of particles in the model.

fn particles_mut(&mut self) -> &mut [Particle]

Return a mutable list of particles in the model.

This is available to allow mutating particles within the model. This should not change the ordering of the particles as these are assumed to be fixed throughout the integration.

fn static_particle_idx<S: AsRef<str>>(
name: S,
i: usize
) -> Result<usize, (S, usize)>

A static implementation of particle_idx which does not need a reference to an instance of the class (i.e. &self).

Errors

If the particle is not within the model, the name and index should be returned as an error so that they can be subsequently handled.

Implementation

If possible, this method should be implemented over particle_idx as models typically will always have the same particle content (with only the value of parameters varying); however, if the particles might vary between instances of the model, this implementation should be unimplemented!() and particle_idx should be implemented instead.

Provided methods

fn hubble_rate(&self, beta: f64) -> f64

Return the Hubble rate at the specified inverse temperature.

The default implementation assumes the Universe to be radiation dominated such that

H(\beta) = \sqrt{\frac{\pi^2}{90}} g_{*}^{1/2}(\beta) \frac{1}{m_{\text{Pl}}} \frac{1}{\beta^2}.

Which is valid provided that the entropy density of the Universe does not change too rapidly.

Warning

The dominated epoch in the Standard Model of cosmology ends at the matter–radiation equality, which occurs at an inverse temperature of $\beta \approx 10^{8}$ GeV^{-1}.

fn len_particles(&self) -> usize

Return the number of particles in the model.

fn particle_idx<S: AsRef<str>>(
&self,
name: S,
i: usize
) -> Result<usize, (S, usize)>

Return the index corresponding to a particle’s name and generation index.

If the particle has no index, the i argument should be ignored. Indices must be 0-indexed, so that generational indices are [0, 1, 2].

Note that since particles are distinguished from anti-particles based on the sign of the integer, particle_idx should not return 0 unless that particle is its own anti-particle (so the distinction is irrelevant).

By default, this implementation defers to static_particle_idx as models generally do not need access to particular instances of the model.

Errors

If the particle is not within the model, the name and index should be returned as an error so that they can be subsequently handled.

fn particle_num<S: AsRef<str>>(
&self,
name: S,
i: usize
) -> Result<isize, (S, usize)>

The signed version of particle_idx.

This is identical to particle_idx but returns an isize type.

Errors

If the particle is not within the model, the name and index should be returned as an error so that they can be subsequently handled.

fn static_particle_num<S: AsRef<str>>(
name: S,
i: usize
) -> Result<isize, (S, usize)>

A static implementation of particle_num which does not need a reference to an instance of the class (i.e. &self).

Errors

If the particle is not within the model, the name and index should be returned as an error so that they can be subsequently handled.

fn particle_name(&self, i: isize) -> Result<String, isize>

Convert a signed particle number to the corresponding particle name.

If the particle number is negative and the particle is not its own antiparticle, a macron (◌̄) will be place over the first character of the name.

Errors

If the particle is not found within the model, the number is returned is it can be handled separately.

fn particle<S: AsRef<str>>(&self, name: S, i: usize) -> &Particle

Return a reference to the matching particle by name.

Panics

Panics if then name or particle is not known within the model.

fn particle_mut(&mut self, name: &str, i: usize) -> &mut Particle

Return a mutable reference to the matching particle by name.

Panics

Panics if then name or particle is not known within the model.

fn as_context(&self) -> Context<'_, Self>

Return a instance of Context for the model.

As this is not within the context of solving the Boltzmann equations, the following attributes have a special value:

step = 0,

step_size = 1.0,

normalization = (hubble_rate * beta * n).recip(),

n is an array of 1.0, and

na is an array of 0.0.

All other attribute contexts will be as expected.