Struct boltzmann_solver::particle::Particle

pub struct Particle {
    pub spin: u8,
    pub mass: f64,
    // some fields omitted
}

Particle type

The particle type is only very basic, containing only the minimum information required for cosmological purposes. Specifically, it only keeps track of the particle's spin, mass and whether it is complex. It does not keep track of transformation properties and interactions.

By default, the degrees of freedom is calculated based on the particle's spin and whether it is complex or not. An arbitrary number can be specified which will override this calculation (which can be used for multiplets for example).

In the long run, it is intended that this type be replaced by another one.

Fields
spin: u8

The spin is stored as twice the spin, so a spin-½ particle has spin == 1 and a spin-1 particle has spin == 2

mass: f64

Mass of the particle in GeV

Methods

impl Particle

pub fn new(spin: u8, mass: f64) -> Self

Create a new particle with the specified spin and mass.

The mass is specified in units of GeV, and the spin is multiplied by a factor of two such that a spin-½ particle has spin == 1 and a spin-1 particle has spin == 2.

pub fn set_complex(self) -> Self

Indicate that the particle is complex.

pub fn set_dof(self, dof: f64) -> Self

Specify how many degrees of freedom this particle has. This overwrites completely the default calculation (that is, the maximum degrees of freedom will be independent of the spin of the particle and whether it is complex or not).

pub fn is_complex(&self) -> bool

Returns true if the particle is complex.

pub fn is_bosonic(&self) -> bool

Returns true if the particle is bosonic.

pub fn is_fermionic(&self) -> bool

Returns true if the particle is fermionic.

pub fn degrees_of_freedom(&self) -> f64

Return the number of degrees of freedom for the underlying particle.

pub fn phase_space(&self, e: f64, mu: f64, beta: f64) -> f64

Return the equilibrium phase space occupation of the particle.

pub fn number_density(&self, mu: f64, beta: f64) -> f64

Return the equilibrium number density of the particle.