Enum boltzmann_solver::prelude::Interaction

pub enum Interaction<M: Model> {
    TwoParticle {
        signed_particles: [isize; 2],
        particles: [usize; 2],
        antiparticles: [f64; 2],
        m2: Box<dyn Fn(&M) -> f64>,
    },
    ThreeParticle {
        signed_particles: [[isize; 3]; 3],
        particles: [[usize; 3]; 3],
        antiparticles: [[f64; 3]; 3],
        m2: Box<dyn Fn(&M) -> f64>,
        asymmetry: Option<Box<dyn Fn(&M) -> f64>>,
    },
    FourParticle {
        signed_particles: [[isize; 4]; 3],
        particles: [[usize; 4]; 3],
        antiparticles: [[f64; 4]; 3],
        m2: Box<dyn Fn(&M, f64, f64, f64) -> f64>,
        gamma: [RwLock<CubicHermiteSpline>; 3],
    },
    Custom {
        signed_particles: [Vec<isize>; 2],
        particles: [Vec<usize>; 2],
        antiparticles: [Vec<f64>; 2],
        gamma: Box<dyn Fn(&M) -> f64>,
    },
}

Interaction between particles.

See the documentation of Interaction::three_particle and Interaction::four_particle for more details as to their implementations.

Variants

TwoParticle

Two particle interaction.

These are mass-mixing interactions such as \(m N N^{c}\) for the right-handed neutrino, or \(\mu_{12}^{2} \Phi_1 \Phi_2\) in the case of a two Higgs doublets. In the limit that the strength of the mixing is very small, then it might be sensible to consider the asymmetry in the densities of the two particles involved.

This is not implemented yet.

Fields of TwoParticle

signed_particles: [isize; 2]

Particles involved.

particles: [usize; 2]

Indices of the particles involved.

antiparticles: [f64; 2]

Antiparticle indicators.

m2: Box<dyn Fn(&M) -> f64>

Squared amplitude as a function of the model.

ThreeParticle

Three particle interaction.

Every permutation of the three particle interaction is computed through crossing symmetry, though only the interaction between the heaviest particle and the lighter two is computed.

Fields of ThreeParticle

signed_particles: [[isize; 3]; 3]

The three permutations of the particles involved.

particles: [[usize; 3]; 3]

The three permutations of the indices of the particles involved.

antiparticles: [[f64; 3]; 3]

The three permutations particle/antiparticle indicators.

m2: Box<dyn Fn(&M) -> f64>

Squared amplitude as a function of the model.

asymmetry: Option<Box<dyn Fn(&M) -> f64>>

Asymmetry between the amplitude and its \(\mathcal{CP}\) conjugate.

FourParticle

Four particle interaction.

Every permutation of the three particle interaction is compute through crossing symmetry.

Fields of FourParticle

signed_particles: [[isize; 4]; 3]

The three permutations of the particles involved.

particles: [[usize; 4]; 3]

The three permutations of the indices of the particles involved.

antiparticles: [[f64; 4]; 3]

The three permutations particle/antiparticle indicators.

m2: Box<dyn Fn(&M, f64, f64, f64) -> f64>

The squared amplitude as a function of the model and the Mandelstam variables s, t and u corresponding to the particle ordering as specified in the call of Interaction::four_particle (which is also the first of the three permutations above).

gamma: [RwLock<CubicHermiteSpline>; 3]

Spline of the interaction rate density which is built as the interaction is used.

Custom

Custom interaction where an arbitrary number of ingoing particle interact with an arbitrary number of outgoing particles.

Unlike all other interactions, this only computes one permutation of the interaction and does not use crossing symmetry. Furthermore, the interaction rate \(\gamma\) must be specified explicitly unlike the other interactions in which \(\gamma\) is calculated from the squared amplitude.

Fields of Custom

signed_particles: [Vec<isize>; 2]

Vector of ingoing and outgoing particles.

particles: [Vec<usize>; 2]

Vector of ingoing and outgoing particle indices.

antiparticles: [Vec<f64>; 2]

Vector of ingoing and outgoing particle/antiparticle indicators.

gamma: Box<dyn Fn(&M) -> f64>

Interaction rate density.

Methods

impl<M: Model> Interaction<M>

pub fn two_particle<F>(m2: F, p1: isize, p2: isize) -> Self where
F: Fn(&M) -> f64 + 'static,

Create a new two-particle interaction.

This will calculate the interaction between the particles \(p_1\) and \(p_2\).

Panics

Panics if p1 == p2. For a mass interaction, make sure that p1 == - p2.

pub fn three_particle<F>(m2: F, p1: isize, p2: isize, p3: isize) -> Self where
F: Fn(&M) -> f64 + 'static,

Create a new three-particle interaction.

This will calculate the interaction between the three particles \(p_1\), \(p_2\) and \(p_3\). The order in which the interaction takes place is determined by the masses of the particles such that the heaviest is (inverse) decaying from/to the lighter two.

Particle are indicated by positive integers and the corresponding antiparticles are indicated by negative integers.

If the process differs from its CP-conjugate process, the asymmetry can be specified with Interaction::asymmetry.

Only the interaction between the heavier particle and the lighter two is computed.

pub fn four_particle<F>(
m2: F,
p1: isize,
p2: isize,
p3: isize,
p4: isize
) -> Self where
F: Fn(&M, f64, f64, f64) -> f64 + 'static,

Add a new four-particle interaction.

These are used to compute the various \(2 \leftrightarrow 2\) scatterings between the particles:

\begin{equation} \begin{aligned} p_1 p_2 \leftrightarrow p_3 p_4, \\ p_1 \overline{p_3} \leftrightarrow \overline{p_2} p_4, \\ p_1 \overline{p_4} \leftrightarrow \overline{p_2} p_3. \\ \end{aligned} \end{equation}

Particles are indicated by positive integers and the corresponding antiparticles are indicates by negative integers.

If the process differs from its CP-conjugate process, the asymmetry can be specified with Interaction::asymmetry.

The squared amplitude is defined of the form

\begin{equation} \abs{\mathcal{M}(p_1 p_2 \leftrightarrow p_3 p_4)}(s, t, u) \end{equation}

where \(s\), \(t\) and \(u\) are the usual Mandelstam variables. Crossing symmetry is then used in order to compute the other processes.

The three-body decays are not currently computed.

pub fn custom<F>(gamma: F, ingoing: Vec<isize>, outgoing: Vec<isize>) -> Self where
F: Fn(&M) -> f64 + 'static,

Create a new custom interaction which takes ingoing particles going to outgoing particles.

Unlike the other implementations, this requires an explicit expression for gamma and does not use any crossing symmetry to compute other interactions.

pub fn is_two_particle(&self) -> bool

Check if this is a two-particle interaction.

pub fn is_three_particle(&self) -> bool

Check if this is a three-particle interaction.

pub fn is_four_particle(&self) -> bool

Check if this is a four-particle interaction.

pub fn is_custom(&self) -> bool

Check if this is a custom interaction.

pub fn asymmetry<F>(self, asym: F) -> Self where
F: Fn(&M) -> f64 + 'static,

Set the asymmetry between this interaction and its CP-conjugate.

For three-body interactions, this is defined as:

\begin{equation} \varepsilon \defeq \abs{\mathcal{M}(p_1 \leftrightarrow p_2 p_3)}^2 - \abs{\mathcal{M}(\overline{p_1} \leftrightarrow \overline{p_2} \overline{p_3})}^2. \end{equation}

If an asymmetry is defined, then the decay of the heavier particle will generate an asymmetry in the lighter two.

Asymmetries for four-particle interactions are currently unsupported.

pub fn particles(&self) -> Vec<Vec<isize>>

Return the particles involved in this interaction.

The resulting vector is a Vec<Vec<isize>> where the outer list are for the different possible interactions (calculated using crossing symmetry), and the inner list as the signed particle numbers involved in the specific interaction.

pub fn particles_idx(&self) -> Vec<Vec<usize>>

Analogue of Interaction::particles, but instead of returning the signed particle numbers, return purely the particle index which does not distinguish between particles and anti-particles.

pub fn gamma(&self, c: &Context<M>) -> Vec<f64>

Calculate the value(s) of gamma.

pub fn width(&self, c: &Context<M>) -> (Vec<isize>, f64)

Calculate the decay width associated with a particular interaction.

This returns two values:

a vector of particles involved, with the first particle being the one decaying and the remaining being the daughter particles; and

the partial width of this process in GeV.

pub fn change(
&self,
dn: Array1<f64>,
dna: Array1<f64>,
c: &Context<M>
) -> (Array1<f64>, Array1<f64>)

Add this interaction to the dn array.

The changes in dn should contain only the effect from the integrated collision operator and not take into account the normalization to inverse-temperature evolution nor the dilation factor from expansion of the Universe. These factors are handled separately and automatically.

This function automatically adjusts the rate so that overshooting is avoided.