Struct boltzmann_solver::prelude::Context

pub struct Context<'a, M> {

    pub step: u64,
    pub substep: i8,
    pub beta: f64,
    pub step_size: f64,
    pub hubble_rate: f64,
    pub normalization: f64,
    pub eq: Array1<f64>,
    pub eqn: Array1<f64>,
    pub n: Array1<f64>,
    pub na: Array1<f64>,
    pub model: &'a M,
    pub in_equilibrium: &'a [usize],
    pub no_asymmetry: &'a [usize],
    // some fields omitted

}

Current context at a particular step in the numerical integration.

Fields
step: u64

Evaluation step.

substep: i8

The substep apply within Runge-Kutta integration and are numbered starting from 0. If no substep is applicable, a negative number (typically -1) is used.

beta: f64

Inverse temperature in GeV$^{-1}$

Note that the value of beta at the next integration step will in general not be beta + step_size as the values of beta for substep fall in between the current and the next values of beta.

step_size: f64

Step size.

Note that the value of beta at the next integration step will in general not be beta + step_size as the values of beta for substep fall in between the current and the next values of beta.

hubble_rate: f64

Hubble rate, in GeV

normalization: f64

Normalization factor, which is

\frac{1}{H \beta n_1}

where $n_1$ is the number density of a single massless bosonic degree of freedom, $H$ is the Hubble rate and $\beta$ is the inverse temperature.

eq: Array1<f64>

Equilibrium number densities for the particles at the current value of beta.

eqn: Array1<f64>

Equilibrium number densities for the particles at the next integration step.

n: Array1<f64>

Current number density

na: Array1<f64>

Current number density asymmetries

model: &'a M

Model data

in_equilibrium: &'a [usize]

List of particles which are forced in equilibrium.

If the context

no_asymmetry: &'a [usize]

List of particles which are forced to never develop any asymmetry.