Struct boltzmann_solver::prelude::Particle

pub struct Particle {
    pub own_antiparticle: bool,
    pub mass: f64,
    pub mass2: f64,
    pub width: f64,
    pub width2: f64,
    pub decays: HashMap<Vec<isize>, f64>,
    pub name: String,
    // some fields omitted
}

Particle type

The particle type is only very basic, containing only the minimum information required for cosmological purposes. Specifically, it only keeps track of the particle's spin, mass and whether it is complex. It does not keep track of transformation properties and interactions.

By default, the degrees of freedom is calculated based on the particle's spin and whether it is complex or not. An arbitrary number can be specified which will override this calculation (which can be used for multiplets for example).

For hashing purposes, only the particle's name is taken into account.

In the long run, it is intended that this type be replaced by another one.

Fields
own_antiparticle: bool

Whether the particle is its own antiparticle or not.

By default, this is assumed to be false and can be set to true with Particle::own_antiparticle.

mass: f64

Mass of the particle in GeV.

This should be updated using Particle::set_mass so that both the mass and squared mass are updated simultaneously.

mass2: f64

Squared mass of the particle in GeV².

This should be updated using Particle::set_mass so that both the mass and squared mass are updated simultaneously.

width: f64

Width of the particle in GeV.

This should be updated using Particle::set_width so that both the width and squared width are updated simultaneously.

width2: f64

Squared width of the particle in GeV².

This should be updated using Particle::set_width so that both the width and squared width are updated simultaneously.

decays: HashMap<Vec<isize>, f64>

Decays

name: String

Name

Implementations

impl Particle

#[must_use]pub fn new(spin: u8, mass: f64, width: f64) -> Self

Create a new particle with the specified spin and mass.

The mass is specified in units of GeV, and the spin is multiplied by a factor of two such that a spin-½ particle has spin == 1 and a spin-1 particle has spin == 2.

pub fn set_mass(&mut self, mass: f64) -> &mut Self

Set the mass of the particle.

pub fn set_width(&mut self, width: f64) -> &mut Self

Set the width of the particle.

#[must_use]pub fn real(self) -> Self

Indicate that the particle is real.

This function returns self and should be used in constructors.

#[must_use]pub fn complex(self) -> Self

Indicate that the particle is complex.

This function returns self and should be used in constructors.

#[must_use]pub fn own_antiparticle(self) -> Self

Indicate that the particle is its own antiparticle, thereby preventing any asymmetry from being generated in its density.

#[must_use]pub fn dof(self, dof: f64) -> Self

Specify how many internal degrees of freedom this particle has.

This is a multiplicative factor to the degrees of freedom. For a 'pseudo' particle such as \(B-L\), this should be set to zero.

This function returns self and should be used in constructors.

#[must_use]pub fn name<S: Into<String>>(self, name: S) -> Self

Specify the particle's name

#[must_use]pub fn is_real(&self) -> bool

Returns true if the particle is real (real scalar, Majorana fermion)

#[must_use]pub fn is_complex(&self) -> bool

Returns true if the particle is complex (complex scalar, Dirac fermion)

#[must_use]pub fn is_bosonic(&self) -> bool

Returns true if the particle is bosonic.

#[must_use]pub fn is_fermionic(&self) -> bool

Returns true if the particle is fermionic.

#[must_use]pub fn degrees_of_freedom(&self) -> f64

Return the number of degrees of freedom for the underlying particle.

The general formula is:

\begin{equation} N_\text{int} \times N_\text{spin} \times N_\text{complex} \end{equation}

where \(N_\text{int}\) are the internal degrees of freedom, \(N_\text{spin}\) are the spin degrees of freedom, and \(N_\text{complex}\) are the degrees of freedom for complex particle.

The spin degrees of freedom are:

spin-0: 1

spin-1/2: 2

spin-1: 2 for massless, 3 for massive

spin-3/2: 4

spin-2: 2 for massless, 5 for massive

or more generally, for half-integer spins n/2, the degrees of freedom is n + 1; and for integer spin n > 0, the degrees of freedom are 2 for massless or 2n + 1 for massive.

#[must_use]pub fn phase_space(&self, e: f64, mu: f64, beta: f64) -> f64

Return the equilibrium phase space occupation of the particle.

#[must_use]pub fn number_density(&self, mu: f64, beta: f64) -> f64

Return the equilibrium number density of the particle.

#[must_use]pub fn normalized_number_density(&self, mu: f64, beta: f64) -> f64

Return the equilibrium number density of the particle, normalized to the number density of a massless boson with one degree of freedom.

#[must_use]pub fn entropy_dof(&self, beta: f64) -> f64

Return the entropy degrees of freedom associated with this particle.