Struct boltzmann_solver::prelude::ParticleData

pub struct ParticleData {
    pub own_antiparticle: bool,
    pub mass: f64,
    pub mass2: f64,
    pub width: f64,
    pub width2: f64,
    pub dof: f64,
    pub name: String,
    /* private fields */
}

Particle type

The particle type is only very basic, containing only the minimum information required for cosmological purposes. Specifically, it only keeps track of the particle’s spin, mass and whether it is complex. It does not keep track of transformation properties and interactions.

By default, the degrees of freedom is calculated based on the particle’s spin and whether it is complex or not. An arbitrary number can be specified which will override this calculation (which can be used for multiplets for example).

For hashing purposes, only the particle’s name is taken into account.

In the long run, it is intended that this type be replaced by another one.

Fields
own_antiparticle: bool

Whether the particle is its own antiparticle or not.

By default, this is assumed to be false and can be set to true with [Particle::own_antiparticle].

mass: f64

Mass of the particle in GeV.

This should be updated using [Particle::set_mass] so that both the mass and squared mass are updated simultaneously.

mass2: f64

Squared mass of the particle in GeV².

This should be updated using [Particle::set_mass] so that both the mass and squared mass are updated simultaneously.

width: f64

Width of the particle in GeV.

This should be updated using [Particle::set_width] so that both the width and squared width are updated simultaneously.

width2: f64

Squared width of the particle in GeV².

This should be updated using [Particle::set_width] so that both the width and squared width are updated simultaneously.

dof: f64

Internal degrees of freedom of the particle

name: String

Name

Implementations

impl Data

pub fn new(representation: LorentzRepresentation, mass: f64, width: f64) -> Self

Create a new particle with the specified spin and mass.

The mass is specified in units of GeV, and the spin is multiplied by a factor of two such that a spin-½ particle has spin == 1 and a spin-1 particle has spin == 2.

pub fn set_mass(&mut self, mass: f64) -> &mut Self

Set the mass of the particle.

pub fn set_width(&mut self, width: f64) -> &mut Self

Set the width of the particle.

pub fn own_antiparticle(self) -> Self

Indicate that the particle is its own antiparticle, thereby preventing any asymmetry from being generated in its density.

pub fn dof(self, dof: f64) -> Self

Specify how many internal degrees of freedom this particle has. These can be because the particle is complex or any other internal gauge symmetries.

This is a multiplicative factor to the degrees of freedom. For a ‘pseudo’ particle such as $B-L$, this should be set to zero.

This function returns self and should be used in constructors.

pub fn name<S: Into<String>>(self, name: S) -> Self

Specify the particle’s name.

Panics

The particle name cannot be an empty string.

pub fn is_bosonic(&self) -> bool

Returns true if the particle is bosonic.

pub fn is_fermionic(&self) -> bool

Returns true if the particle is fermionic.

pub fn is_massless(&self) -> bool

Check whether the particle is massless.

As the calculation for the mass may not be exact, this checks whether the mass square is less than the minimum positive normal floating point (equivalent to $m \lesssim 10^{-150}$).

pub fn degrees_of_freedom(&self) -> f64

Return the number of degrees of freedom for the underlying particle.

The general formula is:

N_\text{int} \times N_\text{spin}

where $N_\text{int}$ are the internal degrees of freedom, $N_\text{spin}$ are the spin degrees of freedom.

The spin degrees of freedom are:

spin-0: 1

spin-1/2: 2

spin-1: 2 for massless, 3 for massive

spin-3/2: 4

spin-2: 2 for massless, 5 for massive

or more generally, for half-integer spins n/2, the degrees of freedom is n + 1; and for integer spin n > 0, the degrees of freedom are 2 for massless or 2n + 1 for massive.

Panics

Panics if a particle with spin greater than 2 is given.

pub fn phase_space(&self, e: f64, mu: f64, beta: f64) -> f64

Return the equilibrium phase space occupation of the particle.

pub fn number_density(&self, beta: f64, mu: f64) -> f64

Return the equilibrium number density of the particle.

pub fn normalized_number_density(&self, beta: f64, mu: f64) -> f64

Return the equilibrium number density of the particle, normalized to the number density of a massless boson with one degree of freedom.

pub fn entropy_dof(&self, beta: f64) -> f64

Return the entropy degrees of freedom associated with this particle.

pub fn propagator<M: Model>(&self, model: &M, s: f64) -> Complex<f64>

Return the propagator denominator for the particle.

The propagator is defined as follow:

\mathcal{P}_i^{-1}(s) = s - m_i^2 + i \theta(s) m_i \Gamma_i,

where $\theta$ is the Heaviside step function.

pub fn ris_propagator<M: Model>(&self, model: &M, s: f64) -> SinglePropagator

Return the RIS propagator for the particle.

The RIS propagator is defined such that a real intermediate state is substracted from the propagator. This is done to avoid a double-counting issue of with $2 \xleftrightarrow{P} 2$ interaction with some intermediate particle $P$ and the combination of the two 3-particle interactions 2 ↔ P and P ↔ 2.

The definition is follows that of a regular propagator except in the following case:

\mathcal{P}_i(s) \mathcal{P}_i^*(s)
= \frac{(s - m_i^2)^2 - (m_i \Gamma_i)^2}{[(s - m_i^2)^2 + (m_i \Gamma_i)^2]^2}.