Trait boltzmann_solver::prelude::Propagator

pub trait Propagator {
    fn conj(self) -> Self;

    fn eval(&self) -> Complex<f64>;

    fn norm_sqr(&self) -> f64;
}

Trait to handle particle proapgators.

This is defined to correctly handle real intermediate states. Specifically, the propagator is defined such that

P_{i}^*(s) P_{j}(s)
= \begin{cases}
  \frac{(s - m_i^2)^2 - (m_i \Gamma_i)^2}
       {[(s - m_i^2)^2 + (m_i \Gamma_i)^2]^2} & i = j \\
  P_i^*(s) P_j(s) & i \neq j
\end{cases}

where

P_{i}^{-1}(s) \defeq s - m_i^2 + i \theta(s) m_i \Gamma_i.

When implementing this, it is always expected that it will be multiplied by complex conjugated propagator at which point the Propagator type will produce a floating point. For example:

use boltzmann_solver::prelude::*;

let p1 = Particle::new(0, 125.0, 1e-3);
let p2 = Particle::new(1, 350.0, 1.0);

let m11 = p1.propagator(125.0) * p1.propagator(125.0).conj();
let m12 = p1.propagator(125.0) * p2.propagator(125.0).conj();
let m22 = p2.propagator(125.0) * p2.propagator(125.0).conj();

let m = (p1.propagator(125.0) + p2.propagator(125.0)) * (p1.propagator(125.0) + p2.propagator(125.0)).conj();

assert!((m11 + m12 + m12.conj() + m22 - m).re.abs() < f64::EPSILON);
assert!((m11 + m12 + m12.conj() + m22 - m).im.abs() < f64::EPSILON);

Required methods

fn conj(self) -> Self

Return the complex conjugate the propagator.

fn eval(&self) -> Complex<f64>

Evaluate the propagator.

This is generally not used explicity as the amplitude itself is rarely evaluated, only the squared amplitude.

fn norm_sqr(&self) -> f64

Compute the norm squared of the propagator.

This is equivalent to having p * p.conj(), but does not require the allocation of a second propagator.