Struct boltzmann_solver::solver::number_density::Interaction

pub struct Interaction { /* fields omitted */ }

Store an interaction.

The interaction consists of two elements: a list of particles involved and the rate at which this reaction is going.

All interactions are always assumed to be of the form the general form

\begin{equation} \left( \frac{n_{\vt i}}{n_{\vt i}^{(0)}} - \frac{n_{\vt f}}{n_{\vt f}^{(0)}} \right) \gamma, \end{equation}

Where each \(n_{\vt i}\) and \(n_{\vt f}\) are initial and final state particles. As an explicit example, the interaction

Interaction::new([Initial(1), Final(2), Final(2)], gamma)

is equivalent to the following interactions:

\begin{equation} \begin{aligned} \ddfrac{n_1}{t} &= -\left( \frac{n_1}{n_1^{(0)}} - \frac{n_2}{n_2^{(0)}} \frac{n_3}{n_3^{(0)}} \right) \gamma, \\ \ddfrac{n_2}{t} &= \left( \frac{n_1}{n_1^{(0)}} - \frac{n_2}{n_2^{(0)}} \frac{n_3}{n_3^{(0)}} \right) \gamma, \\ \ddfrac{n_3}{t} &= \left( \frac{n_1}{n_1^{(0)}} - \frac{n_2}{n_2^{(0)}} \frac{n_3}{n_3^{(0)}} \right) \gamma. \end{aligned} \end{equation}

Methods

impl Interaction

pub fn new<I>(particles: I, rate: f64) -> Self where
I: IntoIterator<Item = Interacting>,

Create a new interaction with the specified particles involved.

pub fn decay(self, m2: f64) -> Self

Specify the amplitude squared for a two-body decay.

This option assumes that there is only one initial particle and that is decaying in a two-body decay. By specifying this, we avoid computing \(\gamma / n^{(0)}\) as both numerator and denominator go to 0 very fast and fail to compute.

The decay rate is computed using:

\begin{equation} \begin{aligned} \frac{n_a}{n_a^{(0)}} \gamma(a \to bc) &= n_a \frac{\abs{\mathcal M}^2}{16 \pi m_a} \frac{K_1(m_a \beta)}{K_2(m_b \beta)}, \\ \frac{Y_a}{Y_a^{(0)}} \gamma(a \to bc) &= Y_a \frac{\abs{\mathcal M}^2}{16 \pi m_a} \frac{K_1(m_a \beta)}{K_2(m_b \beta)} \end{aligned} \end{equation}

and the inverse decay rate is adapted from that.

pub fn dn<M: Model + Sync>(
&mut self,
dn: Array1<f64>,
n: &Array1<f64>,
c: &Context<M>
) -> Array1<f64>

Add this interaction to the dn array.

The changes in dn should contain only the effect from the integrated collision operator and not take into account the normalization to inverse-temperature evolution nor the dilation factor from expansion of the Universe. These factors are handled separately and automatically.

This function automatically adjusts the rate so that overshooting is avoided.