Struct boltzmann_solver::solver::number_density::NumberDensitySolver

pub struct NumberDensitySolver<M: Model> { /* fields omitted */ }

Boltzmann equation solver for the number density.

All number densities are normalized to that of a massless boson with a single degree of freedom (\(g = 1\)). As a result of this convention, \(n_\gamma = 2\) as the photon has two degrees of freedom.