Struct boltzmann_solver::solver::phase_space::Context

pub struct Context {
    pub beta: f64,
    pub hubble_rate: f64,
    pub eq_f: Array2<f64>,
    pub eq_boson: Array1<f64>,
    pub eq_fermion: Array1<f64>,
}

Context provided containing pre-computed values which might be useful when evaluating interactions.

Fields
beta: f64

Inverse temperature in GeV^{-1}

hubble_rate: f64

Hubble rate, in GeV

eq_f: Array2<f64>

Equilibrium phase spaces for the particles. The first axis has the particles and follows the same order as specified to the solver. The second axis has the energies

eq_boson: Array1<f64>

Equilibrium phase space for massless bosons. This is specified per degree of freedom (that is \(g = 1\))

eq_fermion: Array1<f64>

Equilibrium phase space for massless fermions. This is specified per degree of freedom (that is \(g = 1\))