Struct boltzmann_solver::solver::phase_space::PhaseSpaceSolver

pub struct PhaseSpaceSolver { /* fields omitted */ }

The solver holding all the information.

Methods

impl PhaseSpaceSolver

pub fn new() -> Self

Create a new instance of the phase space solver.

The default range of temperatures span 1 GeV through to 10^{20} GeV.

pub fn beta_range(self, start: f64, end: f64) -> Self

Set the range of inverse temperature values over which the phase space is evolved.

Inverse temperature must be provided in units of GeV^{-1}.

This function has a convenience alternative called [PhaseSpaceSolver::temperature_range] allowing for the limits to be specified as temperature in the units of GeV.

Panics

Panics if the starting value is larger than the final value.

pub fn temperature_range(self, start: f64, end: f64) -> Self

Set the range of temperature values over which the phase space is evolved.

Temperature must be provided in units of GeV.

This function is a convenience alternative to [PhaseSpaceSolver::beta_range].

Panics

Panics if the starting value is smaller than the final value.

pub fn energy_steps(self, energy_steps: usize) -> Self

Specify the number of energy steps to use in the energy lattice.

By default, this is set to 2048 steps.

Panic

Panics if the energy steps is zero.

pub fn step_change(self, increase: f64, decrease: f64) -> Self

Specify the granularity of the way time evolution is done.

The time evolution is done with a step size of \(h\) such that \(\beta \to \beta + h\) in the next step. As the range of \(\beta\) spans several orders of magnitude, the step size must be adjusted during the time evolution. This is done by estimating the error at each step and if it falls below a particular threshold, the step size is increased multiplicatively by \(h \to h \times \Delta_{+}\). Similarly, if the estimated error becomes too large, the step size is decreases multiplicatively by \(h \to h \times \Delta_{-}\).

The value of \(\Delta_{+}\) is specified by increase and the value of \(\Delta_{-}\) is specified by decrease. They are 1.1 and 0.5 respectively by default.

Panic

This will panic if the increase factor is not greater than 1.0 or if the decrease factor is not less than 1.0.

pub fn initialize(self) -> Self

Initialize the phase space solver.

pub fn add_species(&mut self, s: Species) -> &mut Self

Add a particle species.

The initial conditions for this particle are generated assuming the particle to be in thermal and chemical equilibrium at the initial temperature.

If the particle and anti-particle are to be treated separately, the two species have to be added.

pub fn solve<U>(&self, universe: &U) -> Array2<f64> where
U: Universe,

Evolve the initial conditions by solving the PDEs.

Note that this is not a true PDE solver, but instead converts the PDE into an ODE in time (or inverse temperature in this case), with the derivative in energy being calculated solely from the previous time-step.