Struct boltzmann_solver::solver::Solver

pub struct Solver<M> { /* fields omitted */ }

Boltzmann solver

Implementations

impl<M> Solver<M> where
M: ModelInteractions,

pub fn solve(&mut self) -> Result<(Array1<f64>, Array1<f64>), Error>

Evolve the initial conditions by solving the PDEs.

Note that this is not a true PDE solver, but instead converts the PDE into an ODE in time (or inverse temperature in this case), with the derivative in energy being calculated solely from the previous time-step.

Errors

If the time evolution encounters an error, the function returns an error. The final number density (despite the error) can still be obtained by using Error::into_inner, though this may not be at the expected final temperature as some errors abort the integration.

pub fn gammas(
&mut self,
size: usize,
normalize: bool
) -> (Vec<String>, Array2<Option<f64>>)

Compute the interaction rates.

The number of logarithmic steps in beta is specified by n, with the range of $\beta$ values being taken from the solver. If normalize is true, the interaction rate is divided by $n_1 H \beta$, where $n_1$ is the equilibrium number density of a single bosonic degree of freedom, $H$ is the Hubble rate and $\beta$ is the inverse temperature.

The interactions rates are returned as two dimensional array with data[[i, j]] giving the interaction rate for the $j$th interaction at the $i$th $\beta$ step. The interaction index corresponds to the index of the interaction as returned by ModelInteractions::interactions, offset by 1 such that the 0th value is for $\beta$ itself.

The entries of the returned array as Option<f64> in order to distinguish cases where the rate is not computed due to being unphysical from cases where it is 0.

pub fn asymmetries(
&mut self,
size: usize,
normalize: bool
) -> (Vec<String>, Array2<Option<f64>>)

Compute the asymmetric interaction rates.

The arguments and returned values are identical to Solver::gammas.