Struct boltzmann_solver::solver_ap::number_density::NumberDensitySolver

pub struct NumberDensitySolver<M: Model> { /* fields omitted */ }

Boltzmann equation solver for the number density.

All number densities are normalized to that of a massless boson with a single degree of freedom (\(g = 1\)). As a result of this convention, \(n_\gamma = 2\) as the photon has two degrees of freedom.

Methods

impl<'a, M: Model> NumberDensitySolver<M>

pub fn threshold_number_density(self, threshold: f64) -> Self

Set the threshold number density to count as 0.

Any number density whose absolute value is less than the threshold will be treated as being exactly zero. This applies to both calculated number densities as well as equilibrium number densities. Furthermore, this also applies to 'abstract' number densities such as \(B-L\).

This is by default set to 0.0.

Panics

Panics if the threshold is negative.