Module boltzmann_solver::statistic

If the rate of collisions between particles is sufficiently high (as is usually the case), the phase space distribution of the particles will quickly converge onto either the Fermi–Dirac statistic or the Bose–Einstein statistic depending on whether the particle is a half-integer or integer spin particle.

Furthermore, in the limit that the occupation is really low, both statistics resemble the Maxwell–Boltzmann distribution in the non-relativistic case, or the Maxwell–Jüttner distribution for the relativistic case.

The statistics are implemented, as well as calculations of the number density.

Enums

Statistic
The statistics which describe the distribution of particles over energy states. Both Fermi–Dirac and Bose–Einstein quantum statistics are implemented, as well as the classical Maxwell–Boltzmann statistic.

Constants

BOSON_EQ_DENSITY
Equilibrium number density for massless bosons, normalized to the equilibrium number density of a massless boson. This is specified per degree of freedom (that is $g = 1$).

FERMION_EQ_DENSITY
Equilibrium number density for massless fermions, normalized to the equilibrium number density of a massless boson. This is specified per degree of freedom (that is $g = 1$)

Traits

Statistics
Equilibrium statistics.