Enum boltzmann_solver::statistic::Statistic

pub enum Statistic {
    FermiDirac,
    BoseEinstein,
    MaxwellBoltzmann,
    MaxwellJuttner,
}

The statistics which describe the distribution of particles over energy states. Both Fermi–Dirac and Bose–Einstein quantum statistics are implemented, as well as the classical Maxwell–Boltzmann statistic.

Variants
FermiDirac

Fermi–Dirac statistic describing half-integer-spin particles:

\begin{equation} f_{\textsc{FD}} = \frac{1}{\exp[(E - \mu) \beta] + 1}. \end{equation}

BoseEinstein

Bose–Einstein statistic describing integer-spin particles:

\begin{equation} f_{\textsc{BE}} = \frac{1}{\exp[(E - \mu) \beta] - 1}. \end{equation}

MaxwellBoltzmann

Maxwell–Boltzmann statistic describing classical particles:

\begin{equation} f_{\textsc{MB}} = \exp[-(E - \mu) \beta]. \end{equation}

MaxwellJuttner

Maxwell–Jüttner statistic describing relativistic classical particles:

\begin{equation} f_{\textsc{MJ}} = \frac{E \beta \sqrt{E^2 - m^2}}{m K_2(m \beta)} \exp[- E \beta]. \end{equation}