Module boltzmann_solver::utilities

Module of various useful miscellaneous functions.

Modules

spline
Cubic Hermite interpolation

Functions

integrate_st
Integrate the amplitude with respect to the Mandelstam variable $s$ and $t$:

kallen_lambda
Kallen lambda function:

kallen_lambda_sqrt
Square root of the Kallen lambda function:

t_range
Return the minimum and maximum value of the Mandelstam variable $t$ based on the four particle squared masses $m_1^2$, $m_2^2$, $m_3^2$ and $m_4^2$, where particles 1 and 2 are initial state and particles 3 and 4 are the final state particles.