Function boltzmann_solver::utilities::kallen_lambda_sqrt

pub fn kallen_lambda_sqrt(a: f64, b: f64, c: f64) -> f64

Square root of the Kallen lambda function:

\begin{equation} \lambda^{\frac{1}{2}}(a, b, c) = \sqrt{a^2 + b^2 + c^2 - 2ab - 2ac - 2bc} \end{equation}

This implementation is more precise than taking the square root of kallen_lambda in cases where the arguments are span several orders of magnitude.

Example

use boltzmann_solver::utilities::{kallen_lambda, kallen_lambda_sqrt};

assert!((kallen_lambda_sqrt(5.0, 2.0, 0.5) - 1.5).abs() < 1e-14);
assert!((kallen_lambda(5.0, 2.0, 0.5).sqrt() - kallen_lambda_sqrt(5.0, 2.0, 0.5)).abs() < 1e-14);
assert!((kallen_lambda_sqrt(1.0, 1.0, 1.0) - 3f64.sqrt()).abs() < 1e-14);

Warning

This function only returns the absolute value of the result.