Struct boltzmann_solver::utilities::spline::Spline

pub struct Spline { /* fields omitted */ }

Linear spline interpolator

Implementations

impl LinearSpline

#[must_use]pub fn empty() -> Self

Create a new empty linear Spline.

#[must_use]pub fn len(&self) -> usize

Return the number of data points in the underlying data.

#[must_use]pub fn is_empty(&self) -> bool

Check whether the spline is empty

pub fn min_points(&mut self, min_points: usize)

Adjust the minimum number of points before accuracy is considered.

pub fn add(&mut self, x: f64, y: f64)

Adds a data point to the spline.

If the x value is already present, the input is ignored and the spline remains unchanged.

If the data point is within the range that is being interpolated, the given value is compared to the interpolated value and if they are in good agreement, both sides adjacent to the new point are marked as accurate (provided the spline has reached the minimum number of points).

Any addition to the spline which extends the interval will result in the new interval being marked as inaccurate.

#[must_use]pub fn accurate(&self, x: f64) -> bool

Check whether a given interval is accurate or not.

If the point checked falls outside of the boundary, this is automatically determined to be false. If the x value is a known control point, then true is returned even if the interval on either side might not be accurate.

#[must_use]pub fn sample(&self, x: f64) -> f64

Sample the spline at the specific x value.

For values of x outside of the domain of the underlying data, the boundary value is returned.