Module boltzmann_solver::model

Model information

Modules

interaction
Common trait and implementations for interactions.

Structs

Particle
Particle type

StandardModel
The Standard Model of particle physics.

Traits

Model
Contains all the information relevant to a particular model, including masses, widths and couplings. All these attributes can be dependent on the inverse temperature \(\beta\).

ModelInteractions
Supertrait for Model for the handling of interactions.