Module boltzmann_solver::model::interaction

Common trait and implementations for interactions.

Structs

FourParticle
Three particle interaction, all determined from the underlying squared amplitude.

InteractionParticleIndices
List of particle indices involved in the interaction.

InteractionParticleSigns
The signed particle to its signum (as a floating point).

InteractionParticles
List of particles involved in the interaction.

PartialWidth
Partial width associated with a single specific interaction.

RateDensity
Partial width from a particle.

ThreeParticle
Three particle interaction, all determined from the underlying squared amplitude.

Traits

Interaction
Generic interaction between particles.

Functions

asymmetry_overshoots
Check whether particle asymmetry i from the model with the given rate change will overshoot 0.

overshoots
Check whether particle i from the model with the given rate change will overshoot equilibrium.