Struct boltzmann_solver::model::interaction::InteractionParticles

pub struct InteractionParticles {
    pub incoming_idx: Vec<usize>,
    pub outgoing_idx: Vec<usize>,
    pub incoming_sign: Vec<f64>,
    pub outgoing_sign: Vec<f64>,
    pub incoming_signed: Vec<isize>,
    pub outgoing_signed: Vec<isize>,
    pub particle_counts: HashMap<usize, (f64, f64)>,
    pub gamma_ratio: Option<f64>,
}

List of particles involved in the interaction.

Fields
incoming_idx: Vec<usize>

Initial state particle indices.

outgoing_idx: Vec<usize>

Final state particle indices.

incoming_sign: Vec<f64>

Initial state particle sign. A positive sign indicates this is a particle, while a negative sign indicates it is an antiparticle.

outgoing_sign: Vec<f64>

Final state particle sign. A positive sign indicates this is a particle, while a negative sign indicates it is an antiparticle.

incoming_signed: Vec<isize>

A combination of incoming_idx and incoming_sign.

outgoing_signed: Vec<isize>

A combination of outgoing_idx and outgoing_sign.

particle_counts: HashMap<usize, (f64, f64)>

Aggregate of the particles to take into account multiplicities.

Incoming particles are taken to have a negative multiplicity (as they are being annihilated) while outgoing particles have a positive multiplicity.

For each particle (indicated by the key), the multiplicity is given ins the tuple for the symmetric and asymmetric rates respectively.

The two counts differ for the symmetric and asymmetric rates because the symmetric rate does not take into account whether it is a particle or antiparticle as both it and its CP-conjugate rate are assumed to be equal. This is evidently not true for the asymmetric rate as it fundamentally assumes that the CP rate is different.

gamma_ratio: Option<f64>

Ratio of the original interactions [Interaction::delta_gamma] to [Interaction::gamma] if defined.

Implementations

impl InteractionParticles

pub fn new(incoming: &[isize], outgoing: &[isize]) -> Self

Create a new set of interaction particles. The absolute value of the numbers indicate the corresponding index for the particle, with the sign indicating whether it is a particle or anti-particle involved.

pub fn cpt(&self) -> Self

Create the CPT conjugate the InteractionParticles. This interchanges the incoming and outgoing particles, and particles with their corresponding antiparticles.

pub fn iter_incoming(&self) -> impl Iterator<Item = (&usize, &f64)>

Iterate over all incoming particle attributes, as a tuple (idx, sign).

pub fn iter_outgoing(&self) -> impl Iterator<Item = (&usize, &f64)>

Iterate over all incoming particle attributes, as a tuple (idx, sign).

pub fn len_incoming(&self) -> usize

Number of incoming particles.

pub fn len_outgoing(&self) -> usize

Number of outgoing particles.

pub fn product_except_incoming(
&self,
arr: &Array1<f64>,
pos: usize,
idx: &[usize]
) -> f64

Compute the product of the relevant entries in the given array as indexed by the incoming particles. This product will exclude the incoming particle given by pos in the list of incoming particles and will also ignore any particle whose index is contained in the idx slice.

For example, if the incoming particles are [1, 3, 2, 1, 4, 2], then the products are:

Result	pos	idx

$n_3 n_2 n_1 n_4 n_2$	0	[]

$n_3 n_1 n_4$	0	[2]

$n_1 n_2 n_1 n_4 n_2$	1	[]

$n_4 n_2$	2	[1, 3]

Warning

The idx slice must be ordered or the result may be wrong.

pub fn product_except_outgoing(
&self,
arr: &Array1<f64>,
i: usize,
idx: &[usize]
) -> f64

Compute the product of the relevant entries in the given array as indexed by the outgoing particles. This product will exclude the outgoing particle given by pos in the list of outgoing particles and will also ignore any particle whose index is contained in the idx slice.

For example, if the outgoing particles are [1, 3, 2, 1, 4, 2], then the products are:

Result	pos	idx

$n_3 n_2 n_1 n_4 n_2$	0	[]

$n_3 n_1 n_4$	0	[2]

$n_1 n_2 n_1 n_4 n_2$	1	[]

$n_4 n_2$	2	[1, 3]

Warning

The idx slice must be ordered or the result may be wrong.

pub fn calculate_particle_counts(&mut self)

Aggregate the particles to take into account multiplicities.

Incoming particles are taken to have a negative multiplicity (as they are being annihilated) while outgoing particles have a positive multiplicity.

For each particle (indicated by the key), the multiplicity is given ins the tuple for the symmetric and asymmetric rates respectively.

The two counts differ for the symmetric and asymmetric rates because the symmetric rate does not take into account whether it is a particle or antiparticle as both it and its CP-conjugate rate are assumed to be equal. This is evidently not true for the asymmetric rate as it fundamentally assumes that the CP rate is different.

pub fn symmetric_prefactor(
&self,
n: &Array1<f64>,
eq: &Array1<f64>,
in_equilibrium: &[usize]
) -> ((f64, f64), (f64, f64))

Net prefactor scaling the interaction rate.

This is defined for a particle $X \to Y$ as

\left( \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right)
- \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right)

If there are some $n_i^{(0)}$ which are zero, then the prefactor will be infinite as the interaction will try and get rid of the particle in question.

If there are multiple particles for which $n_i^{(0)}$ is zero located on both sides, then the direction of the interaction is first determined by the side which contains the most zero equilibrium number densities. If these are also equal, then the result is 0.

\left( \prod_{i \in X} n_i \right) - \left( \prod_{i \in Y} n_i \right)

The result is given as the pair of (numerator, denominator) for incoming and outgoing particles respectively.

Particles which are fixed to be in equilibrium can be specified in the ordered slice in_equilibrium.

Warning

The in_equilibrium slice must be ordered or the result may be wrong.

pub fn symmetric_prefactor_fn<'a>(
&'a self,
n: &'a Array1<f64>,
eq: &'a Array1<f64>,
in_equilibrium: &'a [usize]
) -> impl Fn(f64) -> f64 + 'a

Net prefactor scaling the interaction rate with a specified change included.

The change is calculated by counting the number of times the particles interact and takes into account the multiplicity of the particles involved within the interaction.

Particles which are fixed to be in equilibrium can be specified in the ordered slice in_equilibrium.

Warning

The in_equilibrium slice must be ordered or the result may be wrong.

pub fn symmetric_delta(
&self,
n: &Array1<f64>,
eq: &Array1<f64>,
in_equilibrium: &[usize]
) -> f64

Computes the change in number density required to establish equilibrium.

Given the interaction $X \leftrightarrow Y$, this computes the value of $\delta$ such that

  \prod_{i \in X} \frac{n_i + m_i \delta}{n_i^{(0)}}
= \prod_{i \in Y} \frac{n_i + m_i \delta}{n_i^{(0)}}

holds true up to linear order in $\delta$ and where each $m_i \inZ$ is net number of particle species $i$ involved in the process (-1 and +1 for annihilation and creation respectively).

If one of the equilibrium number density of the incoming particles is 0 then $\delta = \min_{i \in X}\{n_i / m_i\}$ as the interaction can only proceed in one direction. Similarly, if the 0 is on the outgoing side, then $\delta = \min_{i \in Y}\{n_i / m_i}$.

If there are zero equilibrium number densities on both sides, the minimum change required to reach equilibrium is used.

Any particle which is forced to be in equilibrium can be specified in the ordered slice in_equilibrium. Any change to their number densities is ignored (as it would be fixed anyway).

Warning

The in_equilibrium slice must be ordered or the result may be wrong.

pub fn asymmetric_prefactor(
&self,
n: &Array1<f64>,
na: &Array1<f64>,
eq: &Array1<f64>,
in_equilibrium: &[usize],
no_asymmetry: &[usize]
) -> ((f64, f64), (f64, f64))

Compute the prefactor to the symmetric interaction rate which alters the number density asymmetries.

In the forward direction, this is calculated by

\frac{
  \sum_{i \in X} \Delta_i \prod_{j \neq i} n_j
}{
  \prod_{i \in X} n_i^{(0)}
}

and similarly for the backward rate. An example of a $1 \to 2, 3$ rate is:

\frac{\Delta_1}{n_1^{(0)}} - \frac{\Delta_2 n_3 + \Delta_3 n_2}{n_2^{(0)} n_3^{(0)}}

The result is given as the pair of (numerator, denominator) for incoming and outgoing particles respectively.

Particles which are fixed to be in equilibrium can be specified in the ordered slice in_equilibrium, and similarly for no_asymmetry.

Warning

Both the in_equilibrium and no_asymmetry slices must be ordered or the result may be wrong.

pub fn asymmetric_prefactor_fn<'a>(
&'a self,
n: &'a Array1<f64>,
na: &'a Array1<f64>,
eq: &'a Array1<f64>,
in_equilibrium: &'a [usize],
no_asymmetry: &'a [usize]
) -> impl Fn(f64) -> f64 + 'a

Net asymmetric prefactor scaling the interaction rate with a specified change included.

The change is calculated by counting the number of times the particles interact and takes into account the multiplicity of the particles involved within the interaction.

Particles which are fixed to be in equilibrium can be specified in the ordered slice in_equilibrium, and similarly for no_asymmetry.

Warning

Both the in_equilibrium and no_asymmetry slices must be ordered or the result may be wrong.

pub fn asymmetric_delta(
&self,
n: &Array1<f64>,
na: &Array1<f64>,
eq: &Array1<f64>,
in_equilibrium: &[usize],
no_asymmetry: &[usize]
) -> f64

Given the interaction $X \leftrightarrow Y$, this computes the alue of $\delta$ such that

\frac{
  \sum_{i \in X} (\Delta_i + m_i \delta)
  \prod_{\substack{j \in X \\ j \neq i} n_i}
}{
  \prod_{i \in X} n_i^{(0)}
}
=
\frac{
  \sum_{i \in Y} (\Delta_i + m_i \delta)
  \prod_{\substack{j \in Y \\ j \neq i} n_i}
}{
  \prod_{i \in Y} n_i^{(0)}
}

holds true.

If there are zero equilibrium number densities on both sides, the minimum change required to reach equilibrium is used.

Any particle which is forced to be in equilibrium can be specified in the ordered slice in_equilibrium. Any change to their number densities is ignored (as it would be fixed anyway).

Particles which are fixed to never have any asymmetry can be specified in the ordered slice no_asymmetry so they they are ignored for the purposes of this function.

Warning

Both the in_equilibrium and no_asymmetry slices must be ordered or the result may be wrong.

pub fn fast_interaction(
&self,
n: &Array1<f64>,
na: &Array1<f64>,
eq: &Array1<f64>,
in_equilibrium: &[usize],
no_asymmetry: &[usize]
) -> FastInteractionResult

For a fast interaction, compute the change to each particle species such that equilibrium is established.

The computation in the symmetric case is described in Self::symmetric_delta, and the asymmetric case is described in Self::asymmetric_delta.

Warning

Both the in_equilibrium and no_asymmetry slices must be ordered or the result may be wrong.

pub fn display<M>(&self, model: &M) -> Result<String, DisplayError> where
M: Model,

Output a ‘pretty’ version of the interaction particles using the particle names from the model.

Errors

If any particles can’t be found in the model, this will produce an error.

pub fn short_display(&self) -> String

Output a ‘pretty’ version of the interaction particles using the particle indices.

Unlike Self::display, this does not require the model and will always work.