Struct boltzmann_solver::model::interaction::ThreeParticle

pub struct ThreeParticle<M> { /* fields omitted */ }

Three particle interaction, all determined from the underlying squared amplitude.

Implementations

impl<M> ThreeParticle<M> where
M: Model,

pub fn new<F>(squared_amplitude: F, p1: isize, p2: isize, p3: isize) -> Self where
F: Fn(&M) -> f64 + Sync + 'static,

Create a new three-particle interaction.

The order in which the interaction takes place is determined by the masses of the particles such that the heaviest is (inverse) decaying from/to the lighter two.

Particle are indicated by positive integers and the corresponding antiparticles are indicated by negative integers.

If the process differs from its $\CP$-conjugate process, the asymmetry can be specified with ThreeParticle::asymmetry.

pub fn new_all<F>(
squared_amplitude: F,
p1: isize,
p2: isize,
p3: isize
) -> Vec<Self> where
F: Fn(&M) -> f64 + Sync + Copy + 'static,

Create the set of related three particle interactions.

The interactions are all related through crossing symmetry:

$p_1 \leftrightarrow p_2 p_3$,

$\overline{p_2} \leftrightarrow \overline{p_1} p_3$, and

$\overline{p_3} \leftrightarrow \overline{p_1} p_2$.

pub fn set_asymmetry<F>(&mut self, asymmetry: F) -> &Self where
F: Fn(&M) -> f64 + Sync + 'static,

Specify the asymmetry between this process and its $\CP$-conjugate.

This asymmetry is specified in terms of the asymmetry in the squared amplitudes:

\delta \abs{\scM}^2
  \defeq \abs{\scM(p_1 \to p_2 p_3)}^2 - \abs{\scM(\overline{p_1} \to \overline{p_2} \overline{p_3})}^2
  = \abs{\scM(p_1 \to p_2 p_3)}^2 - \abs{\scM(p_2 p_3 \to p_1)}^2

This asymmetry is subsequently used to compute the asymmetric interaction rate given by Interaction::delta_gamma.

pub fn enable_width(self, v: bool) -> Self

Adjust whether this interaction will calculate decay widths.

pub fn enable_gamma(self, v: bool) -> Self

Adjust whether this interaction will calculate decay widths.