Trait boltzmann_solver::model::interaction::Interaction

pub trait Interaction<M: Model> {
    pub fn particles(&self) -> &InteractionParticles;

    pub fn width_enabled(&self) -> bool;

    pub fn gamma_enabled(&self) -> bool;

    pub fn gamma(&self, c: &Context<'_, M>, real: bool) -> Option<f64>;

    pub fn width(&self, c: &Context<'_, M>) -> Option<PartialWidth> { ... }

    pub fn delta_gamma(&self, c: &Context<'_, M>, real: bool) -> Option<f64> { ... }

    pub fn symmetric_prefactor(&self, c: &Context<'_, M>) -> f64 { ... }

    pub fn asymmetric_prefactor(&self, c: &Context<'_, M>) -> f64 { ... }

    pub fn rate(&self, c: &Context<'_, M>) -> Option<RateDensity> { ... }

    pub fn adjusted_rate(&self, c: &Context<'_, M>) -> Option<RateDensity> { ... }

    pub fn change(
        &self, 
        dn: &mut Array1<f64>, 
        dna: &mut Array1<f64>, 
        c: &Context<'_, M>
    ) { ... }
}

Generic interaction between particles.

Required methods

pub fn particles(&self) -> &InteractionParticles

Return the particles involved in this interaction

pub fn width_enabled(&self) -> bool

Whether this interaction is to be used to determine decays.

If width calcuilations are disable, then Interaction::width is expected to return None all the time. If it is enabled, then Interaction::width is expected to return None only when the decay is not kinematically allowed.

pub fn gamma_enabled(&self) -> bool

Whether this interaction is to be used within the Boltzmann equations.

If this returns true, then Interaction::gamma is expected to return None.

pub fn gamma(&self, c: &Context<'_, M>, real: bool) -> Option<f64>

Calculate the reaction rate density of this interaction.

This must return the result before it is normalized by the Hubble rate or other factors related to the number densities of particles involved. It also must not be normalized to the integration step size. Specifically, this corresponds to $\gamma$ in the following expression:

H \beta n_1 \pfrac{n_a}{\beta} =
  - \left( \frac{n_a}{n_a^{(0)}} \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right) \gamma(a X \to Y)
  + \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right) \gamma(Y \to a X)

Care must be taken to obey Interaction::gamma_enabled in order to avoid unnecessary computations. Specifically, this should always return None when self.gamma_enabled() == true.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the 'real' interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

Loading content...
Provided methods

pub fn width(&self, c: &Context<'_, M>) -> Option<PartialWidth>

Calculate the decay width associated with a particular interaction.

There may not be a result if the decay is not kinematically allowed, or not relevant.

The default implementation simply returns None and must be implemented manually. Care must be taken to obey Interaction::width_enabled in order to avoid unnecessary computation (though the incorrect implementation will not be detrimental other than in performance).

pub fn delta_gamma(&self, c: &Context<'_, M>, real: bool) -> Option<f64>

Asymmetry between the interaction and its $\CP$ conjugate:

\delta\gamma
  \defeq \gamma(\vt a \to \vt b) - \gamma(\overline{\vt a} \to \overline{\vt b})
  = \gamma(\vt a \to \vt b) - \gamma(\vt b \to \vt a)

If there is no (relevant) asymmetry, then this should return None.

Note that his is not the same as the asymmetry specified in creating the interaction, with the latter being defined as the asymmetry in the squared amplitudes and the former being subsequently computed.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the 'real' interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

pub fn symmetric_prefactor(&self, c: &Context<'_, M>) -> f64

Net prefactor scaling the interaction rate.

This is defined for a particle $X \to Y$ as

\left( \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right)
- \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right)

If there are some $n_i^{(0)}$ which are zero, then the prefactor will be infinite as the interaction will try and get rid of the particle in question.

If there are multiple particles for which $n_i^{(0)}$ is zero located on both sides, then the direction of the interaction is first determined by the side which contains the most zero equilibrium number densities. If these are also equal, then the result is 0.

\left( \prod_{i \in X} n_i \right) - \left( \prod_{i \in Y} n_i \right)

pub fn asymmetric_prefactor(&self, c: &Context<'_, M>) -> f64

Compute the prefactor to the symmetric interaction rate which alters the number density asymmetries.

In the forward direction, this is calculated by

\frac{
  \sum_{i \in X} \Delta_i \prod_{j \neq i} n_j
}{
  \prod_{i \in X} n_i^{(0)}
}

and similarly for the backward rate. An example of a $1 \to 2, 3$ rate is:

\frac{\Delta_1}{n_1^{(0)}} - \frac{\Delta_2 n_3 + \Delta_3 n_2}{n_2^{(0)} n_3^{(0)}}

pub fn rate(&self, c: &Context<'_, M>) -> Option<RateDensity>

Calculate the actual interaction rates density taking into account factors related to the number densities of particles involved.

The result is normalized by the Hubble rate, but does not include factors relating to the integration step size. Specifically, it corresponds to the right hand side of the equation:

\pfrac{n_a}{\beta} = \frac{1}{H \beta n_1} \left[
  - \left( \frac{n_a}{n_a^{(0)}} \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right) \gamma(a X \to Y)
  + \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right) \gamma(Y \to a X)
\right].

The normalization factor $(H \beta n_1)^{-1}$ is accessible within the current context through Context::normalization.

The rate density is defined such that the change initial state particles is proportional to the negative of the rates contained, while the change for final state particles is proportional to the rates themselves.

The default implementation uses the output of Interaction::gamma and [Interaction::asymmetry] in order to computer the actual rate.

pub fn adjusted_rate(&self, c: &Context<'_, M>) -> Option<RateDensity>

Compute the adjusted rate to handle possible overshoots of the equilibrium number density.

The output of this function should be in principle the same as Interaction::rate, but may be smaller if the interaction rate would overshoot an equilibrium number density. Furthermore, it is scale by the current integration step size $h$:

\Delta n = \pfrac{n}{\beta} h.

The default implementation uses the output of Interaction::rate in order to computer the actual rate.

This method should generally not be implemented. If it is implemented separately, one must take care to take into account the integration step size which is available in Context::step_size.

pub fn change(
&self,
dn: &mut Array1<f64>,
dna: &mut Array1<f64>,
c: &Context<'_, M>
)

Add this interaction to the dn and dna array.

This must use the final $\Delta n$ taking into account all normalization ($(H \beta n_1)^{-1}$) and numerical integration factors (the step size $h$).

The default implementation uses the output of Interaction::adjusted_rate in order to computer the actual rate.

This method should generally not be implemented.

Loading content...