Trait boltzmann_solver::model::interaction::Interaction

pub trait Interaction<M: Model> {

    fn particles(&self) -> &Particles;

    fn width_enabled(&self) -> bool;

    fn gamma_enabled(&self) -> bool;

    fn gamma(&self, context: &Context<'_, M>, real: bool) -> Option<f64>;

    fn width(&self, context: &Context<'_, M>) -> Option<PartialWidth> { ... }

    fn delta_gamma(&self, context: &Context<'_, M>, real: bool) -> Option<f64> { ... }

    fn symmetric_prefactor(&self, context: &Context<'_, M>) -> f64 { ... }

    fn asymmetric_prefactor(&self, context: &Context<'_, M>) -> f64 { ... }

    fn rate(&self, context: &Context<'_, M>) -> Option<RateDensity> { ... }

    fn is_fast_check(&self, context: &Context<'_, M>) -> bool { ... }

    fn is_fast(&self, rate: &RateDensity, context: &Context<'_, M>) -> bool { ... }

    fn adjusted_rate(&self, context: &Context<'_, M>) -> Option<RateDensity> { ... }

    fn apply_change(
        &self, 
        dn: &mut Array1<f64>, 
        dna: &mut Array1<f64>, 
        context: &Context<'_, M>
    ) { ... }

    fn display(&self, model: &M) -> Result<String, DisplayError> { ... }

}

Generic interaction between particles.

Required methods

fn particles(&self) -> &Particles

Return the particles involved in this interaction

fn width_enabled(&self) -> bool

Whether this interaction is to be used to determine decays.

If width calcuilations are disable, then Interaction::width is expected to return None all the time. If it is enabled, then Interaction::width is expected to return None only when the decay is not kinematically allowed.

fn gamma_enabled(&self) -> bool

Whether this interaction is to be used within the Boltzmann equations.

If this returns true, then Interaction::gamma is expected to return None.

fn gamma(&self, context: &Context<'_, M>, real: bool) -> Option<f64>

Calculate the reaction rate density of this interaction.

This must return the result before it is normalized by the Hubble rate or other factors related to the number densities of particles involved. It also must not be normalized to the integration step size. Specifically, this corresponds to $\gamma$ in the following expression:

H \beta n_1 \pfrac{n_a}{\beta} =
  - \left( \frac{n_a}{n_a^{(0)}} \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right) \gamma(a X \to Y)
  + \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right) \gamma(Y \to a X)

Care must be taken to obey Interaction::gamma_enabled in order to avoid unnecessary computations. Specifically, this should always return None when self.gamma_enabled() == true.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the ‘real’ interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

Provided methods

fn width(&self, context: &Context<'_, M>) -> Option<PartialWidth>

Calculate the decay width associated with a particular interaction.

There may not be a result if the decay is not kinematically allowed, or not relevant.

The default implementation simply returns None and must be implemented manually. Care must be taken to obey Interaction::width_enabled in order to avoid unnecessary computation (though the incorrect implementation will not be detrimental other than in performance).

fn delta_gamma(&self, context: &Context<'_, M>, real: bool) -> Option<f64>

Asymmetry between the interaction and its $\CP$ conjugate:

\delta\gamma
  \defeq \gamma(\vt a \to \vt b) - \gamma(\overline{\vt a} \to \overline{\vt b})
  = \gamma(\vt a \to \vt b) - \gamma(\vt b \to \vt a)

If there is no (relevant) asymmetry, then this should return None.

Note that his is not the same as the asymmetry specified in creating the interaction, with the latter being defined as the asymmetry in the squared amplitudes and the former being subsequently computed.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the ‘real’ interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

fn symmetric_prefactor(&self, context: &Context<'_, M>) -> f64

Net prefactor scaling the interaction rate.

The default implementation makes use of Particles::symmetric_prefactor but can be adjusted if the interaction has a different dependence on the number densities.

fn asymmetric_prefactor(&self, context: &Context<'_, M>) -> f64

Prefactor to the symmetric interaction rate which alters the number density asymmetries.

The default implementation makes use of Particles::asymmetric_prefactor but can be adjusted if the interaction has a different dependence on the number densities.

fn rate(&self, context: &Context<'_, M>) -> Option<RateDensity>

Calculate the actual interaction rates density taking into account factors related to the number densities of particles involved.

The result is normalized by the Hubble rate, but does not include factors relating to the integration step size. Specifically, it corresponds to the right hand side of the equation:

\pfrac{n_a}{\beta} = \frac{1}{H \beta n_1} \left[
  - \left( \frac{n_a}{n_a^{(0)}} \prod_{i \in X} \frac{n_i}{n_i^{(0)}} \right) \gamma(a X \to Y)
  + \left( \prod_{i \in Y} \frac{n_i}{n_i^{(0)}} \right) \gamma(Y \to a X)
\right].

The normalization factor $(H \beta n_1)^{-1}$ is accessible within the current context through Context::normalization.

The rate density is defined such that the change initial state particles is proportional to the negative of the rates contained, while the change for final state particles is proportional to the rates themselves.

The default implementation uses the output of Interaction::gamma and Interaction::delta_gamma in order to computer the actual rate.

fn is_fast_check(&self, context: &Context<'_, M>) -> bool

Checks whether the interaction has already been deemed fast.

On substep 0, this will always return false even is [is_fast] will return true. This will also only work is [is_fast] was checked at substep 0.

If fast interactions are disabled, this always returns false.

fn is_fast(&self, rate: &RateDensity, context: &Context<'_, M>) -> bool

Check if the interaction is determined as ‘fast’.

Physically, this occurs when $\ddfrac{N}{t} \propto \gamma / H \gg 1$, or $\ddfrac{n}{\beta} \propto \gamma / H N \gg 1$.

Note that the comparison is only done on the first substep and the result is stored for subsequent substeps in order to avoid unecessarily compute the interaction rate. This means that if the interaction is initially slow and becomes fast at a subsequent substep, this will still return false. For subsequent substeps, this returns the same result as [is_fast_check].

A true result must mean that we are using fast interactions. If fast interactions are disabled, this always returns false.

fn adjusted_rate(&self, context: &Context<'_, M>) -> Option<RateDensity>

Compute the adjusted rate to handle possible overshoots of the equilibrium number density.

The output of this function should be in principle the same as Interaction::rate, but may be smaller if the interaction rate would overshoot an equilibrium number density. Furthermore, it is scale by the current integration step size $h$:

\Delta n = \pfrac{n}{\beta} h.

The default implementation uses the output of Interaction::rate in order to computer the actual rate.

This method should generally not be implemented. If it is implemented separately, one must take care to take into account the integration step size which is available in Context::step_size.

If an interaction is deeemed to be too fast for the main runge-Kutta routine to provide an accurate result, None is returned and the particle’s [interaction::Particles] should be added to the context’s list of fast interactions (if being used). This allows for the fast interaction to be handled separated.

fn apply_change(
&self,
dn: &mut Array1<f64>,
dna: &mut Array1<f64>,
context: &Context<'_, M>
)

Add this interaction to the dn and dna array.

This must use the final $\Delta n$ taking into account all normalization ($(H \beta n_1)^{-1}$) and numerical integration factors (the step size $h$).

The default implementation uses the output of Interaction::adjusted_rate in order to computer the actual rate.

This method should generally not be implemented.

fn display(&self, model: &M) -> Result<String, DisplayError>

Output a ‘pretty’ version of the interaction particles using the particle names from the model.

See [InteractionParticles::display] for more information.