Trait boltzmann_solver::model::interaction::Interaction

pub trait Interaction<M> where
    M: Model, {
    fn particles(&self) -> &InteractionParticles;

    fn particles_idx(&self) -> &InteractionParticleIndices;

    fn particles_sign(&self) -> &InteractionParticleSigns;

    fn width_enabled(&self) -> bool;

    fn gamma_enabled(&self) -> bool;

    fn gamma(&self, c: &Context<M>, real: bool) -> Option<f64>;

    fn width(&self, _c: &Context<M>) -> Option<PartialWidth> { ... }

    fn asymmetry(&self, _c: &Context<M>, _real: bool) -> Option<f64> { ... }

    fn rate(&self, c: &Context<M>) -> Option<RateDensity> { ... }

    fn adjusted_rate(&self, c: &Context<M>) -> Option<RateDensity> { ... }

    fn change(
        &self, 
        dn: &mut Array1<f64>, 
        dna: &mut Array1<f64>, 
        c: &Context<M>
    ) { ... }
}

Generic interaction between particles.

Required methods

fn particles(&self) -> &InteractionParticles

Return the particles involved in this interaction

fn particles_idx(&self) -> &InteractionParticleIndices

Return the particles involved in this interaction, not distinguishing between particles an antiparticles.

fn particles_sign(&self) -> &InteractionParticleSigns

Return the sign of the particles, with +1 for a particle and -1 for an antiparticle.

fn width_enabled(&self) -> bool

Whether this interaction is to be used to determine decays.

If width calcuilations are disable, then Interaction::width is expected to return None all the time. If it is enabled, then Interaction::width is expected to return None only when the decay is not kinematically allowed.

fn gamma_enabled(&self) -> bool

Whether this interaction is to be used within the Boltzmann equations.

If this returns true, then Interaction::gamma is expected to return None.

fn gamma(&self, c: &Context<M>, real: bool) -> Option<f64>

Calculate the reaction rate density of this interaction.

This must return the result before it is normalized by the Hubble rate or other factors related to the number densities of particles involved. It also must not be normalized to the integration step size.

Care must be taken to obey Interaction::gamma_enabled in order to avoid unnecessary computations. Specifically, this should always return None when self.gamma_enabled() == true.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the 'real' interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

Loading content...
Provided methods

fn width(&self, _c: &Context<M>) -> Option<PartialWidth>

Calculate the decay width associated with a particular interaction.

There may not be a result if the decay is not kinematically allowed, or not relevant.

The default implementation simply returns None and must be implemented manually. Care must be taken to obey Interaction::width_enabled in order to avoid unnecessary computation (though the incorrect implementation will not be detrimental other than in performance).

fn asymmetry(&self, _c: &Context<M>, _real: bool) -> Option<f64>

Asymmetry between the interaction and its $\CP$ conjugate:

\delta\gamma
  \defeq \gamma(\vt a \to \vt b) - \gamma(\overline{\vt a} \to \overline{\vt b})
  = \gamma(\vt a \to \vt b) - \gamma(\vt b \to \vt a)

If there is no (relevant) asymmetry, then this should return None.

Note that his is not the same as the asymmetry specified in creating the interaction, with the latter being defined as the asymmetry in the squared amplitudes and the former being subsequently computed.

As there can be some nice mathematical cancellations between the interaction rate and the number density normalizations, the result may not be the 'real' interaction rate and may be normalized by another factor. For example, decays will be normalized by the equilibrium number density of the decaying particle in order to avoid possible 0 / 0 errors. In order to get the real interaction rate, real should be set to true.

fn rate(&self, c: &Context<M>) -> Option<RateDensity>

Calculate the actual interaction rates density taking into account factors related to the number densities of particles involved.

The result must not be normalized by the Hubble rate, nor include the factors relating to the integration step size.

The rate density is defined such that the change initial state particles is proportional to the negative of the rates contained, while the change for final state particles is proportional to the rates themselves.

fn adjusted_rate(&self, c: &Context<M>) -> Option<RateDensity>

Adjust the backward and forward rates such that they do not overshoot the equilibrium number densities.

The rate density going into this function is expected to be the output of Interaction::rate and must not be previously normalized to the Hubble rate nor include factors relating to the integration step size.

The output of this function however will be normalized by the HUbble rate and will factors relating to the numerical integration.

fn change(&self, dn: &mut Array1<f64>, dna: &mut Array1<f64>, c: &Context<M>)

Add this interaction to the dn and dna array.

The changes in dn should contain only the effect from the integrated collision operator and not take into account the normalization to inverse-temperature evolution nor the dilation factor from expansion of the Universe. These factors are handled separately and automatically.

This function automatically adjusts the rate calculated by Interaction::rate so that overshooting is avoided.

Loading content...