Trait boltzmann_solver::model::ModelInteractions

pub trait ModelInteractions where
    Self: Model + Sync, {
    type Item: Interaction<Self> + Sync;
    fn interactions(&self) -> &[Self::Item];
}

Supertrait for Model for the handling of interactions.

Associated Types

type Item: Interaction<Self> + Sync

The underlying interaction type.

If only three-particle interactions are used, then ThreeParticle would be appropriate here; however, if a combination of interactions might be used then it is necessary to us Box<dyn Interaction<Self>>.

Required methods

fn interactions(&self) -> &[Self::Item]

Return an iterator over all interactions in the model.