Trait boltzmann_solver::model::ModelInteractions

pub trait ModelInteractions where
    Self: Model + Sync, {
    type Item: Interaction<Self> + Sync;
    fn interactions(&self) -> &[Self::Item];

    fn update_widths(&mut self) { ... }
}

Supertrait for Model for the handling of interactions.

Associated Types

type Item: Interaction<Self> + Sync

The underlying interaction type.

If only three-particle interactions are used, then ThreeParticle would be appropriate here; however, if a combination of interactions might be used then it is necessary to us Box<dyn Interaction<Self>>.

Loading content...
Required methods

fn interactions(&self) -> &[Self::Item]

Return an iterator over all interactions in the model.

Loading content...
Provided methods

fn update_widths(&mut self)

Calculate the widths of all particles.

This computes the possible decays of all particles given the interactions specified within the model in order to compute the particle's final width.

The information is stored in each particle's Particle::width and the partial widths are stored in Particle::decays.

Loading content...