Module boltzmann_solver::solver

Solvers for the Boltzmann equation, or sets of Boltzmann equations

Modules

number_density
Solver for the number density evolution given by integrating the Boltzmann equation.

phase_space
Solver for the phase-space form of the Boltzmann equation.

Structs

EmptyModel
An empty model containing no couplings, masses, etc. This is can be used for very simple implementations of the Boltzmann solver.

Enums

InitialCondition
Initial conditions for a particle.

Traits

Model
Contains all the information relevant to a particular model, including masses, widths and couplings. All these attributes can be dependent on the inverse temperature \(\beta\).

Solver
Common interface for the Boltzmann equation solvers.