Struct boltzmann_solver::solver::SolverBuilder

pub struct SolverBuilder<M> {
    pub model: Option<M>,
    // some fields omitted
}

Boltzmann solver builder

Fields
model: Option<M>

Inner model

Implementations

impl<M> SolverBuilder<M>

pub fn new() -> Self

Creates a new builder for the Boltzmann solver.

The default range of temperatures span 1 GeV through to 10$^{20}$ GeV, and it uses normalization by default.

Most of the method for the builder are intended to be chained one after the other.

use boltzmann_solver::prelude::*;
use boltzmann_solver::model::StandardModel;

let mut solver_builder: SolverBuilder<StandardModel> = SolverBuilder::new()
    // .logger(..)
    // .initial_densities(..)
    .beta_range(1e-10, 1e-6);

// let solver = solver_builder.build();

pub fn model(self, model: M) -> Self

Set a model function.

Add a function that can modify the model before it is used by the context. This is particularly useful if there is a baseline model with fixed parameters and only certain parameters are changed.

pub fn initial_densities<I>(self, n: I) -> Self where
I: IntoIterator<Item = (usize, f64)>,

Specify initial number densities explicitly.

The specification is given as an iterator of (usize, f64) tuples where the first entry is the index of the particle whose equilibrium is specified.

If unspecified, all number densities are assumed to be in equilibrium to begin with.

Repeated initial conditions are combined together with any duplicates overiding previous calls.

pub fn initial_asymmetries<I>(self, na: I) -> Self where
I: IntoIterator<Item = (usize, f64)>,

Specify initial number density asymmetries explicitly.

The specification is given as an iterator of (usize, f64) tuples where the first entry is the index of the particle whose asymmetry is specified.

If unspecified, all asymmetries are assumed to be 0 to begin with.

Repeated initial conditions are combined together with any duplicates overiding previous calls.

pub fn beta_range(self, start: f64, end: f64) -> Self

Set the range of inverse temperature values over which the solution is calculated.

Inverse temperature must be provided in units of GeV$^{-1}$.

This function has a convenience alternative called SolverBuilder::temperature_range allowing for the limits to be specified as temperature in the units of GeV.

Panics

Panics if the starting value is larger than the final value.

pub fn temperature_range(self, start: f64, end: f64) -> Self

Set the range of temperature values over which the phase space is evolved.

Temperature must be provided in units of GeV.

This function is a convenience alternative to Solver::beta_range.

Panics

Panics if the starting value is smaller than the final value.

pub fn in_equilibrium<I>(self, eq: I) -> Self where
I: IntoIterator<Item = usize>,

Specify the particles which must remain in equilibrium.

Particles which are held in equilibrium will never have their number density deviate from equilibrium. The corresponding asymmetry may deviate from 0 unless this is also pegged by SolverBuilder::no_asymmetry.

These particles are specified by their index.

Multiple calls of this function will combine the results together.

pub fn no_asymmetry<I>(self, na: I) -> Self where
I: IntoIterator<Item = usize>,

Specify the particles which never develop an asymmetry.

These particles will never develop an asymmetry between their particle and antiparticle densities. This does not prevent the number densities of the particle and antiparticles from deviating from their equilibrium.

These particles are specified by their index.

Multiple calls of this function will combine the results together.

pub fn logger<F>(self, f: F) -> Self where
F: Fn(&Context<'_, M>, &Array1<f64>, &Array1<f64>) + 'static,

Set the logger.

The logger provides some insight into the numerical integration ‘black box’. Specifically, it is run at the start of each integration step and has access to the current value as a &Array1, the change from this step as a &Solution, and the current Context at the start of the integration step. As a result, for the first step, the solution will be equal to the initial conditions.

This is useful if one wants to track the evolution of the solutions and log these in a CSV file.

pub fn step_precision(self, min: f64, max: f64) -> Self

Specify how large or small the step size is allowed to become.

The evolution of number densities are discretized in steps of $h$ such that $\beta_{i+1} = \beta_{i} + h$. The algorithm will determine automatically the optimal step size $h$ such that the error is deemed acceptable; however, one may wish to override this to prevent step sizes which are either too large or too small.

The step precision sets the range of allowed values of $h$ in proportion to the current value of $\beta$:

p_\text{min} \beta < h < p_\text{max} \beta

The default values are $p_\text{min} = 10^{-10}$ and $p_\text{max} = 1$.

The relative step precision has a higher priority on the step size than the error. That is, the step size will never be less than $p_\text{min} \beta$ even if this results in a larger local error than desired.

Panics

This will panic if $p_\text{min} \geq p_\text{max}$ or if $p_\text{min}$ is negative.

pub fn error_tolerance(self, tol: f64) -> Self

Specify the local error tolerance.

The algorithm will adjust the evolution step size such that the local error remains less than the specified the error tolerance.

Note that the error is only ever estimated and thus may occasionally be inaccurate. Furthermore, the Solver::step_precision takes precedence and thus if a large minimum step precision is requested, the local error may be larger than the error tolerance.

Panics

Panics if the tolerance is less than or equal to 0.

pub fn precompute(self, v: bool) -> Self

Upon calling SolverBuilder::build, set whether interactions should be precomputed.

By default, interactions are precomputed.

pub fn fast_interaction(self, v: bool) -> Self

Specify whether to use fast interactions.

Warning

This feature is currently experimental and may not produce reliable results.

pub fn abort_when_inaccurate(self, v: bool) -> Self

Specify whether local inaccuracies cause the integration to abort.

impl<M> SolverBuilder<M> where
M: ModelInteractions,

pub fn build(self) -> Result<Solver<M>, Error>

Build the Boltzmann solver.

Errors

This will produce an error if some of the configurations options are deemed to be invalid.