Trait boltzmann_solver::solver::Solver

pub trait Solver {
    type Solution;
    type Context;
    fn new() -> Self;

    fn beta_range(self, start: f64, end: f64) -> Self;

    fn temperature_range(self, start: f64, end: f64) -> Self;

    fn step_change(self, increase: f64, decrease: f64) -> Self;

    fn step_precision(self, min: f64, max: f64) -> Self;

    fn error_tolerance(self, tol: f64) -> Self;

    fn initialize(self) -> Self;

    fn add_particle(&mut self, p: Particle, initial_condition: InitialCondition);

    fn add_particles<I>(&mut self, iter: I)
    where
        I: IntoIterator<Item = Particle>;

    fn add_interaction<F: 'static>(&mut self, int: F) -> &mut Self
    where
        F: Fn(Self::Solution, &Self::Solution, &Self::Context) -> Self::Solution;

    fn set_logger<F: 'static>(&mut self, f: F) -> &mut Self
    where
        F: Fn(&Self::Solution, &Self::Solution, &Self::Context);

    fn solve<U: Universe>(&self, universe: &U) -> Self::Solution;
}

Common interface for the Boltzmann equation solvers.

Associated Types

type Solution

The final solution by the solver. This will typically be an array for values with the ordering corresponding to the way add_particle was invoked.

type Context

Context containing relevant information precomputed by the solver which can be used in the calculation of the interactions.

This is used in order to avoid running possibly time-consuming functions for each interaction.

Loading content...
Required methods

fn new() -> Self

Create a new instance of the solver.

In general, the solver is instantiated as follows:

let solver = Solver::new()
             .temperate_range(1e10, 1e3)
             // Other settings
             .initialize();

fn beta_range(self, start: f64, end: f64) -> Self

Set the range of inverse temperature values over which the phase space is evolved.

Inverse temperature must be provided in units of GeV^{-1}.

This function has a convenience alternative called [Solver::temperature_range] allowing for the limits to be specified as temperature in the units of GeV.

Panics

Panics if the starting value is larger than the final value.

fn temperature_range(self, start: f64, end: f64) -> Self

Set the range of temperature values over which the phase space is evolved.

Temperature must be provided in units of GeV.

This function is a convenience alternative to [Solver::beta_range].

Panics

Panics if the starting value is smaller than the final value.

fn step_change(self, increase: f64, decrease: f64) -> Self

Specify the granularity of the way time evolution is done.

The time evolution is done with a step size of \(h\) such that \(\beta \to \beta + h\) in the next step. As the range of \(\beta\) spans several orders of magnitude, the step size must be adjusted during the time evolution. This is done by estimating the error at each step and if it falls below a particular threshold, the step size is increased multiplicatively by \(h \to h \times \Delta_{+}\). Similarly, if the estimated error becomes too large, the step size is decreases multiplicatively by \(h \to h \times \Delta_{-}\).

The value of \(\Delta_{+}\) is specified by increase and the value of \(\Delta_{-}\) is specified by decrease. They are 1.1 and 0.5 respectively by default.

Panic

This will panic if the increase factor is not greater than 1.0 or if the decrease factor is not less than 1.0.

fn step_precision(self, min: f64, max: f64) -> Self

Specify what the relative size of the step size can be.

The time evolution is done with a step size of \(h\) such that \(\beta \to \beta + h\) is the next step. In general, one wants \(h < \beta\) when solving the Boltzmann equations. The values of min and max specify how big \(h\) can be relative to \(\beta\):

\begin{equation} p_\text{min} \beta < h < p_\text{max} \beta \end{equation}

The default values are min = 1e-6 and max = 1e-2.

The relative step precision has a higher priority on the step size than the error. That is, the step size will never be less than \(p_\text{min} \beta\) even if this results in a larger local error than desired.

Panic

This will panic if min >= max.

fn error_tolerance(self, tol: f64) -> Self

Specify the local error tolerance.

If the error deviates too far from the specified tolerance, the integration step size is adjusted accordingly.

fn initialize(self) -> Self

Initialize the phase space solver.

fn add_particle(&mut self, p: Particle, initial_condition: InitialCondition)

Add a particle species.

The initial conditions for this particle are generated assuming the particle to be in thermal and chemical equilibrium at the initial temperature.

If the particle and anti-particle are to be treated separately, the two species have to be added.

fn add_particles<I>(&mut self, iter: I) where
I: IntoIterator<Item = Particle>,

Add a multiple particles from a vector or slice.

fn add_interaction<F: 'static>(&mut self, int: F) -> &mut Self where
F: Fn(Self::Solution, &Self::Solution, &Self::Context) -> Self::Solution,

Add an interaction.

The interaction is a functional of the solution at a particular inverse temperature. The function is of the following form:

f(sum: Self::Solution, densities: &Self::Solution, beta: f64) -> Self::Solution

The first argument, sum, contains the sum of all interactions and the second argument, densities, contains the values of the various number densities at the specified beta. The sum is moved into the function and is expected to be returned. For example:

fn f(sum: mut Array1<f64>, densities: Array1<f64>, beta: f64) -> Array1<f64> {
    sum[0] += - densities[1] * beta;
    sum[1] += - densities[0] * beta;
    sum
}

fn set_logger<F: 'static>(&mut self, f: F) -> &mut Self where
F: Fn(&Self::Solution, &Self::Solution, &Self::Context),

Set the logger.

The logger provides some insight into the numerical integration 'black box'. Specifically, it is run at the start of each integration step and has access to the current value as a &Solution, the change from this step as a &Solution, and the current Context at the start of the integration step. As a result, for the first step, the solution will be equal to the initial conditions.

This is useful if one wants to track the evolution of the solutions and log these in a CSV file.

fn solve<U: Universe>(&self, universe: &U) -> Self::Solution

Evolve the initial conditions by solving the PDEs.

Note that this is not a true PDE solver, but instead converts the PDE into an ODE in time (or inverse temperature in this case), with the derivative in energy being calculated solely from the previous time-step.

Loading content...