Trait boltzmann_solver::statistic::Statistics

pub trait Statistics {
    fn phase_space(&self, beta: f64, e: f64, m: f64, mu: f64) -> f64;

    fn number_density(&self, beta: f64, mass: f64, mu: f64) -> f64 { ... }

    fn number_density_asymmetry(&self, beta: f64, mass: f64, mu: f64) -> f64 { ... }

    fn normalized_number_density(&self, beta: f64, mass: f64, mu: f64) -> f64 { ... }

    fn normalized_number_density_asymmetry(
        &self, 
        beta: f64, 
        mass: f64, 
        mu: f64
    ) -> f64 { ... }
}

Equilibrium statistics.

Required methods

fn phase_space(&self, beta: f64, e: f64, m: f64, mu: f64) -> f64

Evaluate the phase space distribution, for a given energy, mass, chemical potential and inverse temperature.

Provided methods

fn number_density(&self, beta: f64, mass: f64, mu: f64) -> f64

Return number density for a particle following the specified statistic.

n = \frac{1}{2 \pi^2} \int_{m}^{\infty} f_{i} u \sqrt{u^2 - m^2} \dd u.

The naïve implementation will perform a numerical integration.

Note to Implementors

If an analytic closed form is available for the integral, it should be preferred over the numerical integration.

fn number_density_asymmetry(&self, beta: f64, mass: f64, mu: f64) -> f64

Return number density asymmetry for a particle following the specified statistic.

n = \frac{1}{2 \pi^2} \int_{m}^{\infty} [f_{i}(\mu) - f_{i}(-\mu)] u \sqrt{u^2 - m^2} \dd u.

The naïve implementation will perform a numerical integration.

Note to Implementors

If an analytic closed form is available for the integral, it should be preferred over the numerical integration.

fn normalized_number_density(&self, beta: f64, mass: f64, mu: f64) -> f64

Return number density for a particle following the specified statistic, normalized to the number density of a massless boson with a single degree of freedom.

fn normalized_number_density_asymmetry(
&self,
beta: f64,
mass: f64,
mu: f64
) -> f64

Return number density for a particle following the specified statistic, normalized to the number density of a massless boson with a single degree of freedom.