Module boltzmann_solver::utilities

Module of various useful miscellaneous functions.

Functions

checked_div
Perform a 'checked' division.

checked_div_ap
Perform a 'checked' division.

integrate_st
Integrate the amplitude with respect to the Mandelstam variable \(s\) and \(t\):

kallen_lambda
Kallen lambda function:

t_min_max
Return the minimum and maximum value of the Mandelstam variable \(t\) based on the four particle masses \(m_1\), \(m_2\), \(m_3\) and \(m_4\), where particles 1 and 2 are initial state and particles 3 and 4 are the final state particles.